Feat: restructure the View files
