+++ platforms/macos/client/Package.swift
@@ -58,5 +58,10 @@ let package = Package(
             name: "AhaKeyConfigAgent",
             path: "Sources/Agent"
         ),
+        .testTarget(
+            name: "AhaKeyConfigTests",
+            dependencies: ["AhaKeyConfig"],
+            path: "Tests/AhaKeyConfigTests"
+        ),
     ]
 )
