+++ platforms/macos/client/Sources/Utilities/Agent/AgentManager.swift
