+++ platforms/macos/client/Sources/Utilities/Audio/NativeSpeechTranscriptionService.swift
