+++ platforms/macos/client/Sources/Utilities/OLED/DefaultOLEDAssets.swift
