+++ platforms/macos/client/Sources/Utilities/OLED/OLEDFrameEncoder.swift
