+++ platforms/macos/client/Sources/Utilities/System/DebugSigningFixer.swift
