+++ platforms/macos/client/Sources/Utilities/Voice/VoiceAgentSessionStore.swift
