+++ platforms/macos/client/Sources/Utilities/Voice/VoiceAssistantModel.swift
