--- platforms/macos/client/Sources/Views/AhaKeyStudioView.swift
@@ -1,3776 +1,0 @@
-import AppKit
-import AhaKeyConfigUI
-import SwiftUI
-import UniformTypeIdentifiers
-import VoiceAgent
-
-private enum AhaKeyWorkspaceSection: String, CaseIterable, Identifiable {
-    case workbench
-    case deviceInfo
-    case ai
-    case usageData
-    case account
-
-    var id: String { rawValue }
-
-    var title: String {
-        switch self {
-        case .workbench: return "配置台"
-        case .deviceInfo: return "设备信息"
-        case .ai: return "AI 引擎"
-        case .usageData: return "使用数据"
-        case .account: return "账号管理"
-        }
-    }
-
-    var systemImage: String {
-        switch self {
-        case .workbench: return "house"
-        case .deviceInfo: return "list.bullet.rectangle"
-        case .ai: return "cpu"
-        case .usageData: return "chart.bar.xaxis"
-        case .account: return "person.crop.circle"
-        }
-    }
-}
-
-struct AhaKeyStudioView: View {
-    @ObservedObject var bleManager: AhaKeyBLEManager
-    @Binding private var rootWorkspaceMode: AhaKeyRootWorkspaceMode
-    @StateObject private var voiceRelay = VoiceRelayService.shared
-    @StateObject private var nativeSpeech = NativeSpeechTranscriptionService.shared
-    @StateObject private var voiceAgentSession = VoiceAgentSessionStore.shared
-    @StateObject private var agentManager = AgentManager.shared
-
-    @State private var studioDraft: AhaKeyStudioDraft
-    @State private var lastSyncedDraft: AhaKeyStudioDraft
-    @State private var selectedMode: AhaKeyModeSlot
-    @State private var selectedPart: AhaKeyStudioPart
-    @State private var lightBarPreview: LightBarPreviewState
-    @State private var lastSyncDate: Date?
-    @State private var syncStatusMessage = "修改会先保存在本地，连接设备后再同步。"
-    @State private var isSyncing = false
-    @State private var showsOLEDPlaybackPreview = false
-    @State private var showsVoiceAgentConfiguration = false
-    @State private var selectedWorkspace: AhaKeyWorkspaceSection = .workbench
-    @AppStorage(AhaKeyAppearanceMode.storageKey) private var appearanceModeRaw = AhaKeyAppearanceMode.light.rawValue
-
-    init(
-        bleManager: AhaKeyBLEManager,
-        rootWorkspaceMode: Binding<AhaKeyRootWorkspaceMode> = .constant(.classic)
-    ) {
-        self.bleManager = bleManager
-        _rootWorkspaceMode = rootWorkspaceMode
-        let initialDraft = AhaKeyStudioStore.load() ?? .default
-        // 注意：不要在这里调用 VoiceRelayService.updateRoutes —— SwiftUI 会因 bleManager
-        // 的 @Published 属性（workMode/电量/连接状态等）频繁重建 view，init 会跟着多次执行。
-        // 任何在 init 里调用 updateRoutes 都会重置 functionRelay 的 holdingRoute（按住状态），
-        // 导致微信等"按住说话"过几秒就自动结束。正确入口在下面的 .onAppear。
-        _studioDraft = State(initialValue: initialDraft)
-        _lastSyncedDraft = State(initialValue: initialDraft)
-        let initialMode = AhaKeyModeSlot(rawValue: bleManager.workMode) ?? .mode0
-        _selectedMode = State(initialValue: initialMode)
-        _selectedPart = State(initialValue: .key1)
-        _lightBarPreview = State(initialValue: .aiRunning)
-    }
-
-    var body: some View {
-        VStack(spacing: 0) {
-            topBar
-            Divider()
-            HStack(spacing: 0) {
-                sidebarPane
-                Divider()
-                workspacePane
-            }
-            .background(AhaKeyUI.ColorToken.card)
-            Divider()
-            statusBar
-        }
-        .frame(minWidth: 1280, minHeight: 820)
-        .background(AhaKeyUI.ColorToken.base)
-        .groupBoxStyle(AhaKeyPlainPanelGroupBoxStyle())
-        .onAppear {
-            agentManager.applyStoredBluetoothPreferenceOnLaunch(bleManager: bleManager)
-            voiceRelay.start()
-            nativeSpeech.start()
-            voiceAgentSession.start(keyboardMode: AhaKeyModeSlot(rawValue: bleManager.workMode) ?? .mode0)
-            applyCursorRejectMacroSelfHealIfNeeded()
-            voiceRelay.updateRoutes(from: studioDraft)
-            SwitchStateNotifier.shared.bind(to: bleManager)
-            NotificationCenter.default.post(
-                name: .ahaKeyKeyboardWorkModeChanged,
-                object: nil,
-                userInfo: ["workMode": bleManager.workMode]
-            )
-        }
-        .onChange(of: studioDraft) { _, newValue in
-            AhaKeyStudioStore.save(newValue)
-            voiceRelay.updateRoutes(from: newValue)
-        }
-        // 键盘物理档位变化（BLE 查询/通知上报）→ 自动切到对应 Mode 标签，
-        // 这样 OLED 预览、快捷键草稿、发出去的 updateState 三者一致。
-        .onChange(of: bleManager.workMode) { _, newValue in
-            if let slot = AhaKeyModeSlot(rawValue: newValue), slot != selectedMode {
-                selectedMode = slot
-            }
-            if let slot = AhaKeyModeSlot(rawValue: newValue) {
-                voiceAgentSession.updateKeyboardMode(slot)
-            }
-        }
-        .alert("Agent", isPresented: Binding(
-            get: { agentManager.agentUserAlert != nil },
-            set: { if !$0 { agentManager.agentUserAlert = nil } }
-        )) {
-            Button("好", role: .cancel) {
-                agentManager.agentUserAlert = nil
-            }
-        } message: {
-            Text(agentManager.agentUserAlert ?? "")
-        }
-        .sheet(isPresented: $showsOLEDPlaybackPreview) {
-            OLEDMotionPreviewSheet(
-                modeTitle: selectedMode.title,
-                assetPath: currentModeDraft.oled.localAssetPath
-            )
-        }
-        .sheet(isPresented: $voiceRelay.showsPermissionOnboarding) {
-            VoicePermissionOnboardingSheet(
-                voiceRelay: voiceRelay,
-                nativeSpeech: nativeSpeech
-            )
-        }
-        .sheet(isPresented: $showsVoiceAgentConfiguration) {
-            voiceAgentConfigurationSheet
-        }
-    }
-
-    private var topBar: some View {
-        HStack(spacing: 16) {
-            HStack(spacing: 10) {
-                Text("AK")
-                    .font(.system(size: 16, weight: .bold))
-                    .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                    .frame(width: 40, height: 40)
-                    .background(
-                        RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                            .fill(AhaKeyUI.ColorToken.primary.opacity(0.12))
-                    )
-                VStack(alignment: .leading, spacing: 2) {
-                    Text("AhaKey Studio")
-                        .font(AhaKeyUI.Font.title3)
-                    Text("Native review prototype")
-                        .font(AhaKeyUI.Font.footnote)
-                        .foregroundStyle(.secondary)
-                }
-            }
-            .padding(.trailing, 2)
-
-            HStack(spacing: 8) {
-                infoPill(
-                    title: bleManager.isConnected ? "已连接" : (bleManager.isScanning ? "扫描中" : "未连接"),
-                    subtitle: bleManager.deviceName ?? "等待设备",
-                    accent: bleManager.isConnected ? .green : .orange
-                )
-                infoPill(
-                    title: "电量",
-                    subtitle: bleManager.isConnected ? "\(bleManager.batteryLevel)%" : "—",
-                    accent: .blue
-                )
-                infoPill(
-                    title: "拨杆",
-                    subtitle: currentSwitchTitle,
-                    accent: currentSwitchTitle == "自动批准" ? .mint : .indigo
-                )
-                configurationModeStatus
-            }
-
-            Spacer(minLength: 0)
-
-            rootWorkspacePicker
-
-            if !bleManager.isConnected, agentManager.bluetoothConnectionOwner == .ahaKeyStudio {
-                Button(bleManager.isScanning ? "扫描中…" : "连接设备") {
-                    bleManager.userInitiatedConnect()
-                }
-                .buttonStyle(AhaKeySecondaryButtonStyle())
-                .disabled(bleManager.isScanning)
-            }
-
-            if isEditingConfiguration {
-                Button(syncToKeyboardButtonTitle) {
-                    syncAllModesToDevice(returnToKeyboardControlWhenDone: false)
-                }
-                .buttonStyle(AhaKeySecondaryButtonStyle())
-                .disabled(!canSyncConfiguration)
-                .help(syncToKeyboardButtonHelp)
-            }
-
-            Button(configurationModeButtonTitle) {
-                handleConfigurationModeButton()
-            }
-            .buttonStyle(AhaKeyPrimaryButtonStyle())
-            .disabled(isSyncing || agentManager.isAgentOperationInProgress)
-            .help(configurationModeButtonHelp)
-
-            Button {
-                toggleAppearanceMode()
-            } label: {
-                Image(systemName: appearanceMode.systemImage)
-            }
-            .buttonStyle(AhaKeyIconButtonStyle())
-            .help(appearanceMode.title)
-
-            Menu {
-                Button("恢复当前模式默认值") {
-                    restoreCurrentModeDefaults()
-                }
-                Button("重新连接设备") {
-                    bleManager.disconnect()
-                    bleManager.userInitiatedConnect()
-                }
-                Button("清空 OLED 预览") {
-                    clearCurrentOLED()
-                }
-                Divider()
-                Button("隐藏到后台") {
-                    NSApp.keyWindow?.close()
-                }
-                Button("退出 AhaKey Studio") {
-                    NSApp.terminate(nil)
-                }
-            } label: {
-                Image(systemName: "ellipsis.circle")
-                    .font(.system(size: 18, weight: .medium))
-                    .frame(width: 32, height: 32)
-            }
-            .menuStyle(.button)
-            .buttonStyle(.borderless)
-        }
-        .padding(.horizontal, 18)
-        .padding(.vertical, 10)
-        .background(chromeBarBackground)
-    }
-
-    private var rootWorkspacePicker: some View {
-        Picker("界面", selection: $rootWorkspaceMode) {
-            ForEach(AhaKeyRootWorkspaceMode.allCases) { mode in
-                Text(mode.title).tag(mode)
-            }
-        }
-        .pickerStyle(.segmented)
-        .labelsHidden()
-        .frame(width: 220)
-        .help("切换 IDE 工作台和 Agent 工作台")
-    }
-
-    private var sidebarPane: some View {
-        VStack(spacing: 0) {
-            HStack(spacing: 10) {
-                Text("AK")
-                    .font(.system(size: 14, weight: .bold))
-                    .foregroundStyle(.primary)
-                    .frame(width: 32, height: 32)
-                    .background(
-                        RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                            .fill(AhaKeyUI.ColorToken.control)
-                    )
-                    .overlay(
-                        RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                            .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-                    )
-                Text("AhaKey Stu...")
-                    .font(AhaKeyUI.Font.title3)
-                    .lineLimit(1)
-                Spacer()
-                Text("Pro Trial")
-                    .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                    .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                    .padding(.horizontal, 10)
-                    .padding(.vertical, 5)
-                    .background(Capsule().fill(AhaKeyUI.ColorToken.primary.opacity(0.12)))
-                    .overlay(Capsule().strokeBorder(AhaKeyUI.ColorToken.primary.opacity(0.35), lineWidth: 1))
-            }
-            .padding(.bottom, 12)
-
-            Divider()
-                .padding(.bottom, 12)
-
-            VStack(alignment: .leading, spacing: 8) {
-                ForEach(AhaKeyWorkspaceSection.allCases) { section in
-                    Button {
-                        switchWorkspace(section)
-                    } label: {
-                        HStack(spacing: 10) {
-                            Image(systemName: section.systemImage)
-                                .font(.system(size: 14, weight: .semibold))
-                                .foregroundStyle(selectedWorkspace == section ? Color.primary : Color.secondary)
-                                .frame(width: 28, height: 28)
-                                .background(
-                                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                                        .fill(selectedWorkspace == section ? Color.primary.opacity(0.08) : AhaKeyUI.ColorToken.control)
-                                )
-                            Text(section.title)
-                                .font(AhaKeyUI.Font.subhead.weight(selectedWorkspace == section ? .semibold : .medium))
-                                .foregroundStyle(selectedWorkspace == section ? Color.primary : Color.secondary)
-                            Spacer(minLength: 0)
-                        }
-                        .padding(.horizontal, 10)
-                        .padding(.vertical, 8)
-                        .background(
-                            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                                .fill(selectedWorkspace == section ? Color.primary.opacity(0.055) : Color.clear)
-                        )
-                    }
-                    .buttonStyle(.plain)
-                }
-            }
-
-            Spacer(minLength: 16)
-
-            VStack(alignment: .leading, spacing: 10) {
-                Text("Pro Trial")
-                    .font(AhaKeyUI.Font.caption.weight(.semibold))
-                    .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                    .textCase(.uppercase)
-                Text("在试用结束前升级到 Pro")
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                    .foregroundStyle(.primary)
-                Text("已使用 3 天，共 30 天")
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                ProgressView(value: 0.1)
-                    .controlSize(.small)
-                Text("升级 · 开发中")
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                    .foregroundStyle(.secondary)
-                    .frame(maxWidth: .infinity)
-                    .padding(.vertical, 9)
-                    .background(
-                        Capsule(style: .continuous)
-                            .fill(AhaKeyUI.ColorToken.control.opacity(0.7))
-                    )
-            }
-            .padding(14)
-            .ahaKeySurface(radius: AhaKeyUI.Radius.medium)
-
-            HStack(spacing: 8) {
-                sidebarDockButton("person", help: "账号") {
-                    switchWorkspace(.account)
-                }
-                sidebarDockButton("tray", help: "收件箱")
-                sidebarDockButton("gearshape", help: "设置")
-                sidebarDockButton("questionmark.circle", help: "帮助")
-            }
-            .padding(.top, 12)
-        }
-        .padding(16)
-        .frame(width: 254)
-        .frame(maxHeight: .infinity)
-        .background(AhaKeyUI.ColorToken.card)
-    }
-
-    private func sidebarDockButton(_ systemImage: String, help: String, action: @escaping () -> Void = {}) -> some View {
-        Button(action: action) {
-            Image(systemName: systemImage)
-                .font(.system(size: 13, weight: .semibold))
-                .frame(width: 32, height: 32)
-                .background(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                        .fill(AhaKeyUI.ColorToken.control)
-                )
-                .overlay(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                        .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-                )
-        }
-        .buttonStyle(.plain)
-        .help(help == "账号" ? help : "\(help) · 开发中")
-        .disabled(help != "账号")
-        .opacity(help == "账号" ? 1 : 0.55)
-    }
-
-    private func switchWorkspace(_ section: AhaKeyWorkspaceSection) {
-        guard selectedWorkspace != section else { return }
-        var transaction = Transaction()
-        transaction.disablesAnimations = true
-        transaction.animation = nil
-        withTransaction(transaction) {
-            selectedWorkspace = section
-        }
-    }
-
-    @ViewBuilder
-    private var workspacePane: some View {
-        ZStack {
-            switch selectedWorkspace {
-            case .workbench:
-                workbenchWorkspace
-            case .deviceInfo:
-                deviceInfoWorkspace
-            case .ai:
-                aiWorkspace
-            case .usageData:
-                usageDataWorkspace
-            case .account:
-                accountWorkspace
-            }
-        }
-        .id("\(rootWorkspaceMode.rawValue)-\(selectedWorkspace.rawValue)")
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .transaction { transaction in
-            transaction.animation = nil
-            transaction.disablesAnimations = true
-        }
-    }
-
-    private var workbenchWorkspace: some View {
-        VStack(spacing: 0) {
-            workspaceHeader(
-                title: "配置台",
-                subtitle: "配置键盘模式、语音键、OLED 与灯条。编辑配置时由 AhaKey Studio 临时接管蓝牙。",
-                trailing: {
-                    HStack(spacing: 8) {
-                        Text(dirtyCount == 0 ? "配置已就绪" : "未同步 \(dirtyCount) 项")
-                            .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                            .foregroundStyle(dirtyCount == 0 ? Color.secondary : Color.orange)
-                            .ahaKeyPill(accent: dirtyCount == 0 ? .green : .orange)
-                        Picker("模式", selection: $selectedMode) {
-                            ForEach(AhaKeyModeSlot.allCases) { mode in
-                                Text(mode.name).tag(mode)
-                            }
-                        }
-                        .pickerStyle(.segmented)
-                        .labelsHidden()
-                        .frame(width: 300)
-                    }
-                }
-            )
-            Divider()
-            if selectedMode == .mode2 {
-                VoiceAgentWorkspaceView(
-                    session: voiceAgentSession,
-                    modeEditorHeader: modeEditorHeader,
-                    onOpenConfiguration: openVoiceAgentConfiguration
-                )
-            } else {
-                HStack(spacing: 0) {
-                    canvasPane
-                    Divider()
-                    inspectorPane
-                }
-            }
-        }
-    }
-
-    private var deviceInfoWorkspace: some View {
-        VStack(spacing: 0) {
-            workspaceHeader(
-                title: "设备信息",
-                subtitle: "查看连接、Agent、BLE、拨杆与日志。这里不做键盘映射，映射配置请回到配置台。",
-                trailing: {
-                    Text(agentManager.bluetoothConnectionOwner.title)
-                        .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                        .foregroundStyle(.secondary)
-                        .ahaKeyPill()
-                }
-            )
-            Divider()
-            HStack(spacing: 0) {
-                devicePreviewLogPane
-                Divider()
-                DeviceInfoView(bleManager: bleManager)
-                    .padding(.top, 8)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity)
-                    .background(AhaKeyUI.ColorToken.card.opacity(0.42))
-            }
-        }
-    }
-
-    private var devicePreviewLogPane: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            VStack(alignment: .leading, spacing: 12) {
-                HStack(alignment: .center, spacing: 12) {
-                    Image(systemName: bleManager.isConnected ? "keyboard.fill" : "keyboard")
-                        .font(.system(size: 28, weight: .semibold))
-                        .foregroundStyle(bleManager.isConnected ? AhaKeyUI.ColorToken.primary : Color.secondary)
-                        .frame(width: 54, height: 54)
-                        .background(
-                            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.large, style: .continuous)
-                                .fill(bleManager.isConnected ? AhaKeyUI.ColorToken.primary.opacity(0.12) : AhaKeyUI.ColorToken.control)
-                        )
-                    VStack(alignment: .leading, spacing: 4) {
-                        Text(bleManager.deviceName ?? "AhaKey Keyboard")
-                            .font(AhaKeyUI.Font.title2)
-                            .lineLimit(1)
-                        Text(bleManager.isConnected ? "设备在线 · \(workModeDisplayName)" : "等待连接")
-                            .font(AhaKeyUI.Font.callout)
-                            .foregroundStyle(.secondary)
-                    }
-                    Spacer()
-                    Circle()
-                        .fill(bleManager.isConnected ? Color.green : Color.orange)
-                        .frame(width: 10, height: 10)
-                }
-
-                LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 10), count: 2), spacing: 10) {
-                    deviceStatTile("电量", value: bleManager.isConnected ? "\(bleManager.batteryLevel)%" : "—", systemImage: "battery.75")
-                    deviceStatTile("固件", value: "v\(bleManager.firmwareMainVersion).\(bleManager.firmwareSubVersion)", systemImage: "shippingbox")
-                    deviceStatTile("信号", value: bleManager.isConnected ? "\(bleManager.signalStrength) dBm" : "—", systemImage: "wifi")
-                    deviceStatTile("拨杆", value: switchStateDisplayName, systemImage: "switch.2")
-                }
-            }
-            .padding(16)
-            .ahaKeySurface()
-
-            VStack(alignment: .leading, spacing: 12) {
-                HStack {
-                    Label("通信日志", systemImage: "terminal")
-                        .font(AhaKeyUI.Font.title2)
-                    Spacer()
-                    Button {
-                        copyBLELog()
-                    } label: {
-                        Image(systemName: "doc.on.doc")
-                            .font(.system(size: 13, weight: .semibold))
-                    }
-                    .buttonStyle(.plain)
-                    .help("复制全部通信日志")
-                    Button {
-                        bleManager.clearLog()
-                    } label: {
-                        Image(systemName: "trash")
-                            .font(.system(size: 13, weight: .semibold))
-                    }
-                    .buttonStyle(.plain)
-                    .help("清空通信日志")
-                }
-
-                communicationLogList
-            }
-            .padding(16)
-            .ahaKeySurface()
-            .frame(maxHeight: .infinity)
-        }
-        .padding(24)
-        .frame(width: 430)
-        .frame(maxHeight: .infinity, alignment: .topLeading)
-        .background(AhaKeyUI.ColorToken.canvas.opacity(0.35))
-    }
-
-    private var communicationLogList: some View {
-        ScrollViewReader { proxy in
-            ScrollView {
-                if bleManager.commLog.isEmpty {
-                    VStack(spacing: 10) {
-                        Image(systemName: "waveform.path.ecg.rectangle")
-                            .font(.system(size: 30, weight: .regular))
-                            .foregroundStyle(.tertiary)
-                        Text("暂无通信记录")
-                            .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                        Text("连接、查询状态或探测协议后，BLE 通信会常驻显示在这里。")
-                            .font(AhaKeyUI.Font.caption)
-                            .foregroundStyle(.secondary)
-                            .multilineTextAlignment(.center)
-                    }
-                    .frame(maxWidth: .infinity, minHeight: 240)
-                } else {
-                    LazyVStack(alignment: .leading, spacing: 6) {
-                        ForEach(bleManager.commLog.suffix(120)) { entry in
-                            HStack(alignment: .top, spacing: 8) {
-                                Text(entry.formattedTime)
-                                    .font(.system(.caption2, design: .monospaced))
-                                    .foregroundStyle(.tertiary)
-                                    .frame(width: 74, alignment: .leading)
-                                Text(entry.message)
-                                    .font(.system(.caption2, design: .monospaced))
-                                    .foregroundStyle(entry.isError ? .red : .secondary)
-                                    .textSelection(.enabled)
-                                    .frame(maxWidth: .infinity, alignment: .leading)
-                            }
-                            .id(entry.id)
-                        }
-                    }
-                    .padding(10)
-                }
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .background(
-                RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                    .fill(Color(nsColor: .textBackgroundColor).opacity(0.64))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                    .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-            )
-            .onChange(of: bleManager.commLog.count) { _, _ in
-                if let last = bleManager.commLog.last {
-                    proxy.scrollTo(last.id, anchor: .bottom)
-                }
-            }
-        }
-    }
-
-    private var aiWorkspace: some View {
-        VStack(spacing: 0) {
-            workspaceHeader(
-                title: "AI 引擎",
-                subtitle: "管理本地优先的语音识别、后处理与模型运行状态。",
-                trailing: {
-                    Text("本地离线")
-                        .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                        .foregroundStyle(.secondary)
-                        .ahaKeyPill()
-                }
-            )
-            Divider()
-            VStack(alignment: .leading, spacing: 16) {
-                Text("此页将承载引擎提供方、模型路径、推理参数等设置。当前为本地 SwiftUI 占位实现；请从侧栏进入配置台继续配置键盘与 Inspector 主流程。")
-                    .font(AhaKeyUI.Font.callout)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-                VStack(alignment: .leading, spacing: 10) {
-                    Label("本地优先", systemImage: "internaldrive")
-                    Label("模型参数", systemImage: "slider.horizontal.3")
-                    Label("运行状态", systemImage: "waveform.path.ecg")
-                }
-                .font(AhaKeyUI.Font.subhead)
-                .ahaKeyPanel()
-                Spacer()
-            }
-            .padding(24)
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-            .background(AhaKeyUI.ColorToken.card.opacity(0.42))
-        }
-    }
-
-    private var usageDataWorkspace: some View {
-        VStack(spacing: 0) {
-            workspaceHeader(
-                title: "使用数据",
-                subtitle: "个人本机使用看板。当前展示来自本地状态的轻量统计，不上传云端。",
-                trailing: {
-                    Text("本机数据")
-                        .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                        .foregroundStyle(.secondary)
-                        .ahaKeyPill()
-                }
-            )
-            Divider()
-            ScrollView {
-                VStack(alignment: .leading, spacing: 16) {
-                    LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 14), count: 4), spacing: 14) {
-                        usageMetricCard("今日语音", value: "\(estimatedVoiceSessions)", detail: "次触发", systemImage: "mic")
-                        usageMetricCard("最近写入", value: lastCommittedCharacterCountText, detail: "字符", systemImage: "text.cursor")
-                        usageMetricCard("配置改动", value: "\(dirtyCount)", detail: "待同步项", systemImage: "slider.horizontal.3")
-                        usageMetricCard("连接状态", value: bleManager.isConnected ? "在线" : "离线", detail: bleManager.deviceName ?? "AhaKey", systemImage: "antenna.radiowaves.left.and.right")
-                    }
-
-                    VStack(alignment: .leading, spacing: 12) {
-                        Text("语音输入链路")
-                            .font(AhaKeyUI.Font.title2)
-                        usageStatusRow("语音服务", value: nativeSpeech.statusMessage, ok: nativeSpeech.microphoneGranted && nativeSpeech.speechRecognitionGranted)
-                        usageStatusRow("后台语音键", value: voiceRelay.statusMessage, ok: voiceRelay.inputMonitoringGranted && voiceRelay.accessibilityGranted)
-                        usageStatusRow("设备同步", value: syncStatusMessage, ok: !hasUnsyncedChanges)
-                    }
-                    .padding(16)
-                    .ahaKeySurface()
-
-                    VStack(alignment: .leading, spacing: 10) {
-                        Text("最近一次识别")
-                            .font(AhaKeyUI.Font.title2)
-                        Text(nativeSpeech.lastCommittedText.isEmpty ? "暂无已写入文本。完成一次语音输入后，这里会展示最近写入摘要。" : nativeSpeech.lastCommittedText)
-                            .font(AhaKeyUI.Font.callout)
-                            .foregroundStyle(nativeSpeech.lastCommittedText.isEmpty ? .secondary : .primary)
-                            .textSelection(.enabled)
-                            .frame(maxWidth: .infinity, alignment: .leading)
-                    }
-                    .padding(16)
-                    .ahaKeySurface()
-                }
-                .padding(24)
-            }
-            .background(AhaKeyUI.ColorToken.card.opacity(0.42))
-        }
-    }
-
-    private var accountWorkspace: some View {
-        VStack(spacing: 0) {
-            workspaceHeader(
-                title: "账号管理",
-                subtitle: "管理个人账户、订阅与本机隐私设置。当前为本地账户骨架，不影响设备业务逻辑。",
-                trailing: {
-                    Text("Pro Trial")
-                        .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                        .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                        .ahaKeyPill(accent: AhaKeyUI.ColorToken.primary)
-                }
-            )
-            Divider()
-            ScrollView {
-                VStack(alignment: .leading, spacing: 16) {
-                    HStack(alignment: .center, spacing: 14) {
-                        Text("AK")
-                            .font(.system(size: 22, weight: .bold))
-                            .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                            .frame(width: 56, height: 56)
-                            .background(
-                                RoundedRectangle(cornerRadius: AhaKeyUI.Radius.large, style: .continuous)
-                                    .fill(AhaKeyUI.ColorToken.primary.opacity(0.12))
-                            )
-                        VStack(alignment: .leading, spacing: 4) {
-                            Text("本机用户")
-                                .font(AhaKeyUI.Font.title2)
-                            Text("未登录云账户 · 配置与使用数据保存在本机")
-                                .font(AhaKeyUI.Font.callout)
-                                .foregroundStyle(.secondary)
-                        }
-                        Spacer()
-                        Text("登录 · 开发中")
-                            .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                            .foregroundStyle(.secondary)
-                            .padding(.horizontal, 15)
-                            .padding(.vertical, 9)
-                            .background(
-                                Capsule(style: .continuous)
-                                    .fill(AhaKeyUI.ColorToken.control.opacity(0.7))
-                            )
-                    }
-                    .padding(18)
-                    .ahaKeySurface()
-
-                    LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 14), count: 2), spacing: 14) {
-                        accountSettingCard("订阅", detail: "Pro Trial · 已使用 3 天，共 30 天", systemImage: "creditcard")
-                        accountSettingCard("隐私", detail: "语音转写和使用数据默认保存在本机", systemImage: "lock.shield")
-                        accountSettingCard("设备", detail: bleManager.deviceName ?? "尚未连接设备", systemImage: "keyboard")
-                        accountSettingCard("支持", detail: "帮助、反馈与诊断导出入口预留", systemImage: "questionmark.circle")
-                    }
-                }
-                .padding(24)
-            }
-            .background(AhaKeyUI.ColorToken.card.opacity(0.42))
-        }
-    }
-
-    private func previewCanvasPane(title: String, subtitle: String) -> some View {
-        VStack(alignment: .leading, spacing: 16) {
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(AhaKeyUI.Font.title2)
-                Text(subtitle)
-                    .font(AhaKeyUI.Font.callout)
-                    .foregroundStyle(.secondary)
-            }
-            VStack(alignment: .leading, spacing: 8) {
-                AhaKeyKeyboardCanvasView(
-                    modeDraft: currentModeDraft,
-                    selectedPart: selectedPart,
-                    lightBarPreview: lightBarPreview,
-                    switchTitle: currentSwitchTitle,
-                    dirtyParts: dirtyPartsForCurrentMode(),
-                    onSelect: { _ in }
-                )
-                .aspectRatio(109.0 / 54.0, contentMode: .fit)
-                .allowsHitTesting(false)
-            }
-            .padding(20)
-            .ahaKeySurface()
-            Spacer()
-        }
-        .padding(24)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        .background(AhaKeyUI.ColorToken.canvas.opacity(0.35))
-    }
-
-    private func workspaceHeader<Trailing: View>(
-        title: String,
-        subtitle: String,
-        @ViewBuilder trailing: () -> Trailing
-    ) -> some View {
-        HStack(alignment: .center, spacing: 18) {
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(AhaKeyUI.Font.largeTitle)
-                    .foregroundStyle(.primary)
-                Text(subtitle)
-                    .font(AhaKeyUI.Font.callout)
-                    .foregroundStyle(.secondary)
-                    .lineLimit(2)
-            }
-            Spacer(minLength: 16)
-            trailing()
-        }
-        .padding(.horizontal, 24)
-        .padding(.vertical, 16)
-        .background(chromeBarBackground)
-    }
-
-    private func usageMetricCard(_ title: String, value: String, detail: String, systemImage: String) -> some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack {
-                Image(systemName: systemImage)
-                    .font(.system(size: 15, weight: .semibold))
-                    .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                    .frame(width: 30, height: 30)
-                    .background(
-                        RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                            .fill(AhaKeyUI.ColorToken.primary.opacity(0.12))
-                    )
-                Spacer()
-            }
-            Text(value)
-                .font(.system(size: 24, weight: .semibold))
-                .foregroundStyle(.primary)
-                .lineLimit(1)
-                .minimumScaleFactor(0.75)
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                Text(detail)
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .padding(16)
-        .ahaKeySurface()
-    }
-
-    private func usageStatusRow(_ title: String, value: String, ok: Bool) -> some View {
-        HStack(alignment: .top, spacing: 10) {
-            Circle()
-                .fill(ok ? Color.green : Color.orange)
-                .frame(width: 8, height: 8)
-                .padding(.top, 5)
-            VStack(alignment: .leading, spacing: 3) {
-                Text(title)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                Text(value)
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            Spacer(minLength: 0)
-        }
-    }
-
-    private func accountSettingCard(_ title: String, detail: String, systemImage: String) -> some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(systemName: systemImage)
-                .font(.system(size: 15, weight: .semibold))
-                .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                .frame(width: 30, height: 30)
-                .background(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                        .fill(AhaKeyUI.ColorToken.primary.opacity(0.12))
-                )
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                Text(detail)
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(16)
-        .ahaKeySurface()
-    }
-
-    private func deviceStatTile(_ title: String, value: String, systemImage: String) -> some View {
-        HStack(alignment: .center, spacing: 9) {
-            Image(systemName: systemImage)
-                .font(.system(size: 13, weight: .semibold))
-                .foregroundStyle(AhaKeyUI.ColorToken.primary)
-                .frame(width: 26, height: 26)
-                .background(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                        .fill(AhaKeyUI.ColorToken.primary.opacity(0.10))
-                )
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                Text(value)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.75)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(10)
-        .background(
-            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                .fill(AhaKeyUI.ColorToken.control.opacity(0.58))
-        )
-    }
-
-    private var configurationModeStatus: some View {
-        HStack(spacing: 8) {
-            Circle()
-                .fill(selectedMode == .mode2 ? Color.purple : (isEditingConfiguration ? Color.blue : Color.green))
-                .frame(width: 8, height: 8)
-            VStack(alignment: .leading, spacing: 1) {
-                Text(configurationModeTitle)
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(.primary)
-                Text(configurationModeDetail)
-                    .font(.caption2)
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-        }
-        .ahaKeyPill(accent: isEditingConfiguration ? .blue : .green)
-        .help("日常使用由 Agent 控制键盘；需要改键、OLED 或同步时，进入编辑配置后由 AhaKey Studio 临时接管蓝牙。")
-    }
-
-    private var appearanceMode: AhaKeyAppearanceMode {
-        AhaKeyAppearanceMode(rawValue: appearanceModeRaw) ?? .light
-    }
-
-    private func toggleAppearanceMode() {
-        appearanceModeRaw = appearanceMode.next.rawValue
-    }
-
-    private var canvasPane: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            modeEditorHeader
-
-            VStack(alignment: .leading, spacing: 8) {
-                AhaKeyKeyboardCanvasView(
-                    modeDraft: currentModeDraft,
-                    selectedPart: selectedPart,
-                    lightBarPreview: lightBarPreview,
-                    switchTitle: currentSwitchTitle,
-                    dirtyParts: dirtyPartsForCurrentMode(),
-                    onSelect: { selectedPart = $0 }
-                )
-                .aspectRatio(109.0 / 54.0, contentMode: .fit)
-                .frame(maxWidth: .infinity)
-
-                Text("点按灯条、屏幕、四个按键或拨杆即可进入对应配置。")
-                    .font(.callout)
-                    .foregroundStyle(.secondary)
-                    .padding(.leading, 20)
-            }
-            .frame(maxWidth: .infinity)
-            .padding(20)
-            .ahaKeySurface()
-
-            HStack(spacing: 12) {
-                manualCallout(
-                    title: "主流程",
-                    detail: "默认键盘控制 -> 点编辑配置 -> 修改 -> 同步到设备 -> 返回控制"
-                )
-                manualCallout(
-                    title: "模式切换",
-                    detail: "短按设备按键切换模式，OLED 会先显示描述约 1 秒，再回到该模式动图"
-                )
-            }
-        }
-        .padding(24)
-        .frame(minWidth: 640, maxWidth: .infinity, maxHeight: .infinity)
-        .background(AhaKeyUI.ColorToken.canvas.opacity(0.35))
-    }
-
-    private var modeEditorHeader: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(alignment: .firstTextBaseline, spacing: 16) {
-                Text("设备展示")
-                    .font(AhaKeyUI.Font.title2)
-                Text("65%")
-                    .font(AhaKeyUI.Font.subhead)
-                    .foregroundStyle(.secondary)
-
-                Spacer(minLength: 0)
-                Text(selectedMode.title)
-                    .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                    .foregroundStyle(.secondary)
-                    .ahaKeyPill()
-            }
-
-            Text(selectedMode.guidance)
-                .font(AhaKeyUI.Font.callout)
-                .foregroundStyle(.secondary)
-        }
-    }
-
-    private var inspectorPane: some View {
-        VStack(spacing: 0) {
-            HStack {
-                Text("配置面板")
-                    .font(AhaKeyUI.Font.title2)
-                Spacer()
-                Text("35%")
-                    .font(AhaKeyUI.Font.subhead)
-                    .foregroundStyle(.secondary)
-            }
-            .padding(.horizontal, 18)
-            .padding(.vertical, 14)
-            .background(chromeBarBackground)
-            Divider()
-
-            ScrollView {
-                VStack(alignment: .leading, spacing: 16) {
-                    inspectorHeader
-
-                    Group {
-                        switch selectedPart {
-                        case .key1, .key2, .key3, .key4:
-                            keyInspector
-                        case .oledDisplay:
-                            oledInspector
-                                .disabled(!isEditingConfiguration)
-                        case .lightBar:
-                            lightBarInspector
-                                .disabled(!isEditingConfiguration)
-                        case .toggleSwitch:
-                            switchInspector
-                                .disabled(!isEditingConfiguration)
-                        }
-                    }
-                }
-                .padding(18)
-            }
-
-            Divider()
-            inspectorSyncHint
-            Divider()
-            inspectorFooter
-        }
-        .frame(width: 430)
-        .frame(maxHeight: .infinity)
-        .background(AhaKeyUI.ColorToken.card.opacity(0.42))
-    }
-
-    private var inspectorHeader: some View {
-        EmptyView()
-    }
-
-    private var inspectorSyncHint: some View {
-        HStack(alignment: .top, spacing: 10) {
-            Image(systemName: hasUnsyncedChanges ? "arrow.down.doc.fill" : "checkmark.circle.fill")
-                .foregroundStyle(hasUnsyncedChanges ? Color.orange : Color.green)
-                .frame(width: 24, height: 24)
-
-            VStack(alignment: .leading, spacing: 3) {
-                Text(hasUnsyncedChanges ? "有改动尚未写入键盘" : "当前配置已就绪")
-                    .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                    .foregroundStyle(.primary)
-                Text(hasUnsyncedChanges ? "同步后配置才会写入硬件；退出编辑前建议先保存配置。" : "修改按键、OLED 或描述后，这里会提示待同步状态。")
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(.horizontal, 18)
-        .padding(.vertical, 12)
-        .background(AhaKeyUI.ColorToken.control.opacity(0.45))
-    }
-
-    private var inspectorFooter: some View {
-        HStack(spacing: 10) {
-            Text(isEditingConfiguration ? "编辑模式" : "键盘控制")
-                .font(AhaKeyUI.Font.footnote.weight(.semibold))
-                .foregroundStyle(isEditingConfiguration ? Color.blue : Color.secondary)
-            Spacer(minLength: 0)
-
-            Button("重发当前模式") {
-                resendCurrentModeToDevice()
-            }
-            .buttonStyle(AhaKeySecondaryButtonStyle())
-            .disabled(!bleManager.isConnected || isSyncing)
-
-            Button(configurationModeButtonTitle) {
-                handleConfigurationModeButton()
-            }
-            .buttonStyle(AhaKeyPrimaryButtonStyle())
-            .disabled(isSyncing || agentManager.isAgentOperationInProgress)
-        }
-        .padding(.horizontal, 18)
-        .padding(.vertical, 12)
-        .background(chromeBarBackground)
-    }
-
-    private var keyInspector: some View {
-        let key = currentSelectedKey
-        return VStack(alignment: .leading, spacing: 16) {
-            keyHeroHeader(for: key)
-            keySummaryStrip(for: key)
-            keyConfigurationCard(for: key)
-            keyPermissionCard(for: key)
-            keyUsageDisclosure(for: key)
-        }
-    }
-
-    private func keyHeroHeader(for key: AhaKeyKeyDraft) -> some View {
-        HStack(alignment: .center, spacing: 14) {
-            Image(systemName: key.role.systemImage)
-                .font(.system(size: 26, weight: .semibold))
-                .foregroundStyle(.primary)
-                .frame(width: 58, height: 58)
-                .background(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.large, style: .continuous)
-                        .fill(AhaKeyUI.ColorToken.control)
-                )
-                .overlay(
-                    RoundedRectangle(cornerRadius: AhaKeyUI.Radius.large, style: .continuous)
-                        .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-                )
-            VStack(alignment: .leading, spacing: 4) {
-                HStack(spacing: 8) {
-                    Text("Key \(key.role.rawValue + 1)")
-                        .font(.system(size: 26, weight: .semibold))
-                        .lineLimit(1)
-                    if partIsDirty(key.role.part) {
-                        Text("未同步")
-                            .font(AhaKeyUI.Font.caption.weight(.semibold))
-                            .foregroundStyle(Color.orange)
-                            .ahaKeyPill(accent: .orange)
-                    }
-                }
-                Text(key.role.title)
-                    .font(AhaKeyUI.Font.callout)
-                    .foregroundStyle(.secondary)
-            }
-            Spacer(minLength: 0)
-        }
-    }
-
-    private func keySummaryStrip(for key: AhaKeyKeyDraft) -> some View {
-        HStack(spacing: 0) {
-            keySummaryItem(
-                icon: key.role == .voice ? "waveform" : key.role.systemImage,
-                accent: .blue,
-                title: "当前功能",
-                value: key.role == .voice ? "语音输入" : key.role.title
-            )
-            Divider().padding(.vertical, 10)
-            keySummaryItem(
-                icon: "square.stack.3d.up",
-                accent: .purple,
-                title: "当前方案",
-                value: key.displaySummary
-            )
-            Divider().padding(.vertical, 10)
-            keySummaryItem(
-                icon: "checkmark",
-                accent: keyAvailabilityColor(for: key),
-                title: "当前状态",
-                value: keyAvailabilityTitle(for: key)
-            )
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 16)
-        .ahaKeySurface(radius: AhaKeyUI.Radius.large)
-    }
-
-    private func keySummaryItem(icon: String, accent: Color, title: String, value: String) -> some View {
-        HStack(spacing: 12) {
-            Image(systemName: icon)
-                .font(.system(size: 18, weight: .semibold))
-                .foregroundStyle(accent)
-                .frame(width: 42, height: 42)
-                .background(Circle().fill(accent.opacity(0.14)))
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(AhaKeyUI.Font.caption.weight(.semibold))
-                    .foregroundStyle(.secondary)
-                Text(value)
-                    .font(AhaKeyUI.Font.title3)
-                    .foregroundStyle(.primary)
-                    .lineLimit(2)
-                    .minimumScaleFactor(0.78)
-            }
-            Spacer(minLength: 4)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-
-    private func keyConfigurationCard(for key: AhaKeyKeyDraft) -> some View {
-        VStack(alignment: .leading, spacing: 16) {
-            HStack {
-                Text("按键配置")
-                    .font(AhaKeyUI.Font.title2)
-                Spacer()
-                Button(keyConfigSyncButtonTitle) {
-                    syncAllModesToDevice(returnToKeyboardControlWhenDone: false)
-                }
-                .buttonStyle(AhaKeyPrimaryButtonStyle())
-                .disabled(!canSyncConfiguration)
-                .help(syncToKeyboardButtonHelp)
-            }
-
-            VStack(spacing: 12) {
-                if key.role == .voice {
-                    inspectorFormRow(title: "功能类型") {
-                        Text("语音输入")
-                            .font(AhaKeyUI.Font.callout)
-                    }
-                    inspectorFormRow(title: "语音方案") {
-                        Picker("语音方案", selection: selectedVoicePresetBinding) {
-                            ForEach(VoicePreset.allCases) { preset in
-                                Text(preset.availableInV1 ? preset.title : "\(preset.title) · 开发中").tag(preset)
-                            }
-                        }
-                        .pickerStyle(.menu)
-                        .labelsHidden()
-                    }
-                    inspectorFormRow(title: "触发方式") {
-                        Text("单击开始，再按结束")
-                            .font(AhaKeyUI.Font.callout)
-                            .foregroundStyle(.secondary)
-                    }
-                    inspectorFormRow(title: "绑定键位") {
-                        Text((key.voicePreset ?? .custom).defaultBinding.displayLabel)
-                            .font(AhaKeyUI.Font.callout)
-                            .foregroundStyle(.secondary)
-                    }
-                } else {
-                    inspectorFormRow(title: "功能类型") {
-                        Text(key.role.title)
-                            .font(AhaKeyUI.Font.callout)
-                    }
-                    inspectorFormRow(title: "绑定模式") {
-                        Picker("绑定模式", selection: selectedKeyBindingModeBinding) {
-                            Text("单键 / 组合键").tag(KeyBindingMode.shortcut)
-                            Text("宏").tag(KeyBindingMode.macro)
-                        }
-                        .pickerStyle(.segmented)
-                        .labelsHidden()
-                    }
-                    if key.usesMacro {
-                        macroEditor(for: key)
-                            .padding(.top, 2)
-                    } else {
-                        inspectorFormRow(title: "绑定键位") {
-                            ShortcutBindingEditor(shortcut: selectedKeyShortcutBinding)
-                        }
-                    }
-                }
-
-                inspectorFormRow(title: "按键名称") {
-                    TextField("例如 Record / Accept / Reject / Enter", text: selectedKeyDescriptionBinding)
-                        .textFieldStyle(.plain)
-                        .font(AhaKeyUI.Font.callout)
-                        .padding(.horizontal, 10)
-                        .frame(height: 34)
-                        .background(
-                            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                                .fill(Color(nsColor: .textBackgroundColor).opacity(0.72))
-                        )
-                        .overlay(
-                            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.small, style: .continuous)
-                                .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-                        )
-                }
-            }
-
-            if key.role == .voice {
-                Text(voicePresetDetail)
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            if currentSelectedKey.description.containsNonASCII {
-                Text("设备 OLED 只稳定支持 ASCII。中文、emoji 和全角字符会在写入时被自动过滤，当前写入：\(currentSelectedKeySanitizedDescription.isEmpty ? "空白" : currentSelectedKeySanitizedDescription)")
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.orange)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-            if !isEditingConfiguration {
-                Text("当前为键盘控制模式，配置项只读。点击顶部“编辑配置”后可修改并同步到设备。")
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-        .padding(16)
-        .ahaKeySurface(radius: AhaKeyUI.Radius.large)
-        .disabled(!isEditingConfiguration)
-    }
-
-    private func inspectorFormRow<Content: View>(title: String, @ViewBuilder content: () -> Content) -> some View {
-        HStack(alignment: .center, spacing: 12) {
-            Text(title)
-                .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                .foregroundStyle(.primary)
-                .frame(width: 82, alignment: .leading)
-            content()
-                .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .frame(minHeight: 40)
-    }
-
-    private func keyPermissionCard(for key: AhaKeyKeyDraft) -> some View {
-        VStack(alignment: .leading, spacing: 14) {
-            Text("状态与权限")
-                .font(AhaKeyUI.Font.title2)
-
-            LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 10), count: 2), spacing: 10) {
-                if key.role == .voice {
-                    permissionTile("后台监听", granted: voiceRelay.isListening || voiceRelay.inputMonitoringGranted)
-                    permissionTile("麦克风", granted: nativeSpeech.microphoneGranted)
-                    permissionTile("输入监控", granted: voiceRelay.inputMonitoringGranted)
-                    permissionTile("语音转写", granted: nativeSpeech.speechRecognitionGranted)
-                } else {
-                    permissionTile("配置草稿", granted: !partIsDirty(key.role.part))
-                    permissionTile("设备连接", granted: bleManager.isConnected)
-                    permissionTile("编辑模式", granted: isEditingConfiguration)
-                    permissionTile("可同步", granted: canSyncConfiguration || !hasUnsyncedChanges)
-                }
-            }
-
-            HStack(spacing: 10) {
-                Button {
-                    voiceRelay.refreshPermissions(deferredTCCRequery: true)
-                    nativeSpeech.refreshPermissions(deferredTCCRequery: true)
-                } label: {
-                    Label("检查权限", systemImage: "checkmark.shield")
-                }
-                .buttonStyle(AhaKeySecondaryButtonStyle())
-
-                Button {
-                    if key.role == .voice {
-                        nativeSpeech.toggleRecordingFromVoiceKey()
-                    } else {
-                        resendCurrentModeToDevice()
-                    }
-                } label: {
-                    Label(key.role == .voice ? "测试按键" : "重发当前模式", systemImage: key.role == .voice ? "waveform" : "arrow.clockwise")
-                }
-                .buttonStyle(AhaKeySecondaryButtonStyle())
-                .disabled(key.role != .voice && (!bleManager.isConnected || isSyncing))
-            }
-        }
-        .padding(16)
-        .ahaKeySurface(radius: AhaKeyUI.Radius.large)
-    }
-
-    private func permissionTile(_ title: String, granted: Bool) -> some View {
-        HStack(spacing: 8) {
-            Circle()
-                .fill(granted ? Color.green : Color.orange)
-                .frame(width: 8, height: 8)
-            VStack(alignment: .leading, spacing: 2) {
-                Text(title)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-                Text(granted ? "已开启" : "未开启")
-                    .font(AhaKeyUI.Font.caption)
-                    .foregroundStyle(.secondary)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(12)
-        .background(
-            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                .fill(AhaKeyUI.ColorToken.control.opacity(0.58))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: AhaKeyUI.Radius.medium, style: .continuous)
-                .strokeBorder(AhaKeyUI.ColorToken.border, lineWidth: 1)
-        )
-    }
-
-    private func keyUsageDisclosure(for key: AhaKeyKeyDraft) -> some View {
-        DisclosureGroup {
-            VStack(alignment: .leading, spacing: 8) {
-                if key.role == .voice {
-                    Text("保持 AhaKey Studio 在后台运行，按实体语音键开始 macOS 原生转写，再按一次结束并写回当前光标。")
-                    Text("若权限显示异常，请在系统设置中开启输入监控、辅助功能、麦克风和语音转写后，完全退出并重新打开本应用。")
-                } else {
-                    Text(key.role.manualText)
-                    Text("完成配置后点击顶部或本卡片内的“同步到设备”，设备侧短按切换模式时会显示按键描述。")
-                }
-            }
-            .font(AhaKeyUI.Font.callout)
-            .foregroundStyle(.secondary)
-            .padding(.top, 8)
-        } label: {
-            HStack(spacing: 10) {
-                Image(systemName: "questionmark.circle")
-                    .foregroundStyle(.secondary)
-                Text("使用说明")
-                    .font(AhaKeyUI.Font.title3)
-                Spacer()
-            }
-        }
-        .padding(16)
-        .ahaKeySurface(radius: AhaKeyUI.Radius.large)
-    }
-
-    private func keyAvailabilityColor(for key: AhaKeyKeyDraft) -> Color {
-        keyAvailabilityTitle(for: key) == "可用" ? .green : .orange
-    }
-
-    private func keyAvailabilityTitle(for key: AhaKeyKeyDraft) -> String {
-        if key.role == .voice {
-            return nativeSpeech.microphoneGranted && nativeSpeech.speechRecognitionGranted ? "可用" : "待授权"
-        }
-        return isEditingConfiguration ? "可编辑" : "只读"
-    }
-
-    // MARK: - 宏编辑器视图
-
-    @ViewBuilder
-    private func macroEditor(for key: AhaKeyKeyDraft) -> some View {
-        let stepCount = key.macro.count
-        let byteCount = stepCount * 2
-        let overLimit = byteCount > 98 // 固件 payload 上限
-
-        VStack(alignment: .leading, spacing: 10) {
-            HStack(alignment: .firstTextBaseline) {
-                Text("步骤（依次执行）")
-                    .font(.callout.weight(.semibold))
-                Spacer()
-                Text("\(stepCount) 步 · \(byteCount) / 98 字节")
-                    .font(.caption)
-                    .foregroundStyle(overLimit ? .red : .secondary)
-            }
-
-            if key.macro.isEmpty {
-                Text("空宏。点下方“添加步骤”开始录制。")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            } else {
-                VStack(spacing: 6) {
-                    ForEach(Array(key.macro.enumerated()), id: \.element.id) { index, step in
-                        macroStepRow(
-                            index: index,
-                            step: step,
-                            totalCount: key.macro.count
-                        )
-                    }
-                }
-            }
-
-            HStack(spacing: 10) {
-                Button {
-                    appendMacroStep()
-                } label: {
-                    Label("添加步骤", systemImage: "plus.circle.fill")
-                }
-                .buttonStyle(.borderedProminent)
-                .disabled(overLimit)
-
-                Button(role: .destructive) {
-                    updateSelectedKey { $0.macro = [] }
-                } label: {
-                    Label("清空", systemImage: "trash")
-                }
-                .buttonStyle(.bordered)
-                .disabled(key.macro.isEmpty)
-            }
-
-            if overLimit {
-                Text("超过固件单键宏 98 字节 / 49 步上限，同步时会被拒绝。")
-                    .font(.caption)
-                    .foregroundStyle(.red)
-            }
-
-            Text("固件按顺序串行发送；延时单位 3ms（最大 765ms）。需要更长延时请叠加多个延时步骤。")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-
-            if !key.macro.isEmpty {
-                Text("预览：\(key.macro.displaySummary)")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(.secondary)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func macroStepRow(index: Int, step: MacroStep, totalCount: Int) -> some View {
-        HStack(spacing: 8) {
-            Text("\(index + 1).")
-                .font(.caption.monospacedDigit())
-                .foregroundStyle(.secondary)
-                .frame(width: 24, alignment: .trailing)
-
-            Picker("", selection: macroStepActionBinding(id: step.id)) {
-                ForEach(MacroAction.allCases) { action in
-                    Text(action.title).tag(action)
-                }
-            }
-            .pickerStyle(.menu)
-            .labelsHidden()
-            .frame(width: 96)
-
-            if step.action.takesKeycodeParam {
-                Picker("", selection: macroStepKeycodeBinding(id: step.id)) {
-                    Text("未设置").tag(UInt8(0))
-                    ForEach(HIDUsage.allOptions, id: \.code) { option in
-                        Text(option.name).tag(option.code)
-                    }
-                }
-                .pickerStyle(.menu)
-                .labelsHidden()
-                .frame(minWidth: 96)
-            } else if step.action.takesDelayParam {
-                // 勿对带标题的 Stepper 用 labelsHidden()，否则连「15 ms」一并被藏掉。
-                HStack(spacing: 8) {
-                    Text("\(max(1, Int(step.param)) * 3) ms")
-                        .font(.caption.monospacedDigit())
-                        .foregroundStyle(.primary)
-                        .frame(minWidth: 44, alignment: .trailing)
-                    Stepper(
-                        "",
-                        value: macroStepDelayBinding(id: step.id),
-                        in: 1...255
-                    )
-                    .labelsHidden()
-                }
-                .frame(minWidth: 120)
-            } else {
-                Color.clear.frame(minWidth: 96, maxHeight: 1)
-            }
-
-            Spacer(minLength: 0)
-
-            Button {
-                moveMacroStep(from: index, by: -1)
-            } label: {
-                Image(systemName: "arrow.up")
-            }
-            .buttonStyle(.borderless)
-            .disabled(index == 0)
-
-            Button {
-                moveMacroStep(from: index, by: 1)
-            } label: {
-                Image(systemName: "arrow.down")
-            }
-            .buttonStyle(.borderless)
-            .disabled(index >= totalCount - 1)
-
-            Button(role: .destructive) {
-                removeMacroStep(at: index)
-            } label: {
-                Image(systemName: "xmark.circle.fill")
-                    .foregroundStyle(.red)
-            }
-            .buttonStyle(.borderless)
-        }
-        .padding(.vertical, 4)
-        .padding(.horizontal, 6)
-        .background(
-            RoundedRectangle(cornerRadius: 8)
-                .fill(Color(nsColor: .controlBackgroundColor).opacity(0.6))
-        )
-    }
-
-    private func macroStepActionBinding(id: UUID) -> Binding<MacroAction> {
-        Binding(
-            get: {
-                currentSelectedKey.macro.first { $0.id == id }?.action ?? .noOp
-            },
-            set: { newAction in
-                updateMacroStep(id: id) { step in
-                    let previous = step.action
-                    step.action = newAction
-                    // 动作换类别后清零 param，避免把 "Enter 的 HID 码 0x28" 当成延时值 ×3ms 解读。
-                    if previous.takesKeycodeParam != newAction.takesKeycodeParam
-                        || previous.takesDelayParam != newAction.takesDelayParam
-                    {
-                        switch newAction {
-                        case .delay:
-                            step.param = 5 // 默认 15ms，比较通用
-                        case .downKey, .upKey:
-                            step.param = HIDUsage.enter
-                        case .noOp, .upAllKeys:
-                            step.param = 0
-                        }
-                    }
-                }
-            }
-        )
-    }
-
-    private func macroStepKeycodeBinding(id: UUID) -> Binding<UInt8> {
-        Binding(
-            get: {
-                currentSelectedKey.macro.first { $0.id == id }?.param ?? 0
-            },
-            set: { newValue in
-                updateMacroStep(id: id) { $0.param = newValue }
-            }
-        )
-    }
-
-    private func macroStepDelayBinding(id: UUID) -> Binding<UInt8> {
-        Binding(
-            get: {
-                currentSelectedKey.macro.first { $0.id == id }?.param ?? 0
-            },
-            set: { newValue in
-                updateMacroStep(id: id) { $0.param = newValue }
-            }
-        )
-    }
-
-    private var oledInspector: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            GroupBox("当前模式的 OLED 动图") {
-                VStack(alignment: .leading, spacing: 14) {
-                    ZStack {
-                        RoundedRectangle(cornerRadius: 12)
-                            .fill(Color.black.opacity(0.9))
-                            .frame(height: 140)
-
-                        if let image = currentOLEDPreviewImage {
-                            Image(nsImage: image)
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .frame(height: 112)
-                                .clipShape(RoundedRectangle(cornerRadius: 8))
-                        } else {
-                            VStack(spacing: 10) {
-                                Image(systemName: "photo.artframe")
-                                    .font(.system(size: 28))
-                                    .foregroundStyle(.white.opacity(0.8))
-                                Text("当前仅支持动图")
-                                    .foregroundStyle(.white.opacity(0.85))
-                                Text("文字、token、模型状态显示开发中")
-                                    .font(.caption)
-                                    .foregroundStyle(.white.opacity(0.55))
-                            }
-                        }
-                    }
-
-                    HStack(spacing: 10) {
-                        Button("选择动图") {
-                            selectOLEDGIF()
-                        }
-                        .buttonStyle(.bordered)
-
-                        Button("预览动图") {
-                            showsOLEDPlaybackPreview = true
-                        }
-                        .buttonStyle(.bordered)
-                        .disabled(currentModeDraft.oled.localAssetPath == nil)
-
-                        Button(bleManager.isUploadingOLED ? "上传中…" : "上传到 \(selectedMode.title)") {
-                            uploadCurrentOLEDToDevice()
-                        }
-                        .buttonStyle(.borderedProminent)
-                        .disabled(!bleManager.isConnected || bleManager.isUploadingOLED || currentModeDraft.oled.localAssetPath == nil)
-
-                        Button("清空") {
-                            clearCurrentOLED()
-                        }
-                        .buttonStyle(.bordered)
-
-                        Spacer()
-
-                        Text("当前目标：\(selectedMode.title)")
-                            .font(.caption)
-                            .foregroundStyle(.secondary)
-                    }
-
-                    Stepper(value: oledFramesPerSecondBinding, in: 5 ... 20) {
-                        Text("播放速度 \(currentModeDraft.oled.framesPerSecond) FPS")
-                    }
-
-                    if let progress = bleManager.oledUploadProgress, bleManager.isUploadingOLED {
-                        VStack(alignment: .leading, spacing: 6) {
-                            ProgressView(value: progress.fractionCompleted)
-                            Text("已写入 \(progress.completedFrames)/\(progress.totalFrames) 帧，分块 \(progress.completedChunks)/\(progress.totalChunks)")
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                        }
-                    }
-
-                    Text("硬性限制：GIF 源文件 ≤ 2 MB。建议 5–20 FPS、最多 74 帧；将自动缩放到 160×80（RGB565）。")
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-
-                    Text(currentModeDraft.oled.statusLine)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-                .padding(.top, 4)
-            }
-
-            GroupBox("显示逻辑") {
-                VStack(alignment: .leading, spacing: 8) {
-                    Text("切换到当前模式时，OLED 会先显示该模式的按键描述，约 1 秒后回到该模式动图。")
-                    Text("后续会继续增加文字状态、token 用量、模型环境等信息显示能力。")
-                }
-                .font(.callout)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var lightBarReadOnlyInfo: some View {
-        HStack(alignment: .top, spacing: 10) {
-            Image(systemName: "info.circle.fill")
-                .foregroundStyle(.blue)
-            VStack(alignment: .leading, spacing: 6) {
-                Text("出厂灯条映射（只读）")
-                    .font(.subheadline.weight(.semibold))
-                Text("灯条由键盘固件根据 Hook 上报的 IDE 状态点亮，本软件不能改写。下表为各业务场景对应的典型 Hook 状态与出厂灯效说明；画布与「预览到设备」均按此表展示。")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 10)
-                .fill(Color.blue.opacity(0.08))
-        )
-    }
-
-    private var lightBarInspector: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            lightBarReadOnlyInfo
-
-            GroupBox("业务状态 → Hook 状态 → 出厂灯效") {
-                VStack(alignment: .leading, spacing: 12) {
-                    let cases = Array(LightBarPreviewState.allCases)
-                    ForEach(Array(cases.enumerated()), id: \.offset) { index, state in
-                        let hw = AhaKeyLightBarDraft.hardwareEffect(for: state)
-                        VStack(alignment: .leading, spacing: 6) {
-                            HStack(alignment: .firstTextBaseline) {
-                                Text(state.title)
-                                    .font(.callout.weight(.semibold))
-                                Spacer(minLength: 8)
-                                Text(hw.title)
-                                    .font(.callout)
-                                    .foregroundStyle(.primary)
-                                    .multilineTextAlignment(.trailing)
-                            }
-                            Text("Hook 上报：\(state.ideState.label)")
-                                .font(.caption)
-                                .foregroundStyle(.tertiary)
-                            Text(hw.detail)
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                        }
-                        if index < cases.count - 1 {
-                            Divider()
-                        }
-                    }
-                }
-                .padding(.top, 4)
-            }
-
-            GroupBox("状态预览") {
-                VStack(alignment: .leading, spacing: 12) {
-                    HStack {
-                        Text(lightBarPreview.title)
-                            .font(.system(.title3, design: .rounded).weight(.semibold))
-                        Spacer()
-                        Button("预览到设备") {
-                            previewCurrentLightEffectOnDevice()
-                        }
-                        .buttonStyle(.borderedProminent)
-                        .disabled(!bleManager.isConnected || currentDevicePreviewIDEState == nil)
-                    }
-
-                    Text("当前画布预览：\(currentLightEffect.title)")
-                        .font(.callout)
-                        .foregroundStyle(.secondary)
-                    Text(currentLightPreviewHint)
-                        .font(.caption)
-                        .foregroundStyle(currentDevicePreviewIDEState == nil ? .orange : .secondary)
-                    if bleManager.isConnected && bleManager.workMode != 0 {
-                        Text("ℹ️ 出厂固件只在物理 Mode 0（1、2 灯）下完整映射了 state → 灯效；当前键盘在 Mode \(bleManager.workMode)，点预览多半看不到效果，把拨杆切到 Mode 0 再试。")
-                            .font(.caption)
-                            .foregroundStyle(.orange)
-                    }
-
-                    LazyVGrid(columns: [GridItem(.adaptive(minimum: 130), spacing: 10)], spacing: 10) {
-                        ForEach(LightBarPreviewState.allCases) { state in
-                            Button {
-                                lightBarPreview = state
-                            } label: {
-                                VStack(alignment: .leading, spacing: 6) {
-                                    Text(state.title)
-                                        .font(.callout.weight(.semibold))
-                                    Text(AhaKeyLightBarDraft.hardwareEffect(for: state).title)
-                                        .font(.caption)
-                                        .foregroundStyle(.secondary)
-                                        .lineLimit(2)
-                                }
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .padding(12)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 12)
-                                        .fill(state == lightBarPreview ? Color.accentColor.opacity(0.14) : Color(nsColor: .controlBackgroundColor))
-                                )
-                                .overlay(
-                                    RoundedRectangle(cornerRadius: 12)
-                                        .stroke(state == lightBarPreview ? Color.accentColor : Color.black.opacity(0.08), lineWidth: 1)
-                                )
-                            }
-                            .buttonStyle(.plain)
-                        }
-                    }
-                }
-                .padding(.top, 4)
-            }
-
-            GroupBox("说明") {
-                VStack(alignment: .leading, spacing: 8) {
-                    Text("若需让某套灯效出现，请用下方「预览到设备」向固件发送一次对应可达的 IDE 状态试灯；或在 Agent 连上键盘后，通过实际触发 Hook 观察。")
-                }
-                .font(.callout)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-        }
-    }
-
-    private var switchInspector: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            GroupBox("实时档位") {
-                VStack(alignment: .leading, spacing: 12) {
-                    HStack {
-                        Text(currentSwitchTitle)
-                            .font(.system(.title3, design: .rounded).weight(.semibold))
-                        Spacer()
-                        Circle()
-                            .fill(currentSwitchTitle == "自动批准" ? Color.green : Color.indigo)
-                            .frame(width: 10, height: 10)
-                    }
-                    Text("拨杆是物理档位，不是按下瞬态。0 档显示“自动批准”，1 档显示“手动批准”。这里只读取键盘上报的位置，不模拟物理拨动。")
-                        .font(.callout)
-                        .foregroundStyle(.secondary)
-                }
-                .padding(.top, 4)
-            }
-
-            switchEffectivenessBox
-
-            if bleManager.switchState == 0 {
-                GroupBox {
-                    VStack(alignment: .leading, spacing: 8) {
-                        Label("自动批准依赖 Agent 与 Hook，且须蓝牙由 Agent 占用", systemImage: "exclamationmark.triangle.fill")
-                            .foregroundStyle(.orange)
-                            .font(.callout.weight(.semibold))
-                        Text("Claude：PermissionRequest 时返回 allow。Cursor：preToolUse（以及你自建的 beforeShell beforeMCP 钩子）stdout 会返回 permission=allow。若 agent 未连键盘或本 App 占着 BLE，会退成交回确认。须 Agent 在跑、Hook 已装。涉及 shell/删文件等高危时建议用手动档。需要逐条确认时请把拨杆切到“手动批准”。")
-                            .font(.callout)
-                            .foregroundStyle(.secondary)
-                    }
-                }
-            }
-
-            GroupBox("如何理解这个部件") {
-                VStack(alignment: .leading, spacing: 8) {
-                    Text("自动批准：Claude 用 PermissionRequest；Cursor 用 preToolUse 等并返回 JSON permission。均需 Agent + Hooks，且设备蓝牙由 Agent 连接。")
-                    Text("手动批准：会交回用户/终端确认。若 Cursor 仍弹窗，请看 diagnostics 里 ide=cursor 与 diagnostic 字段。")
-                    Text("若仍出现手动：在「设备信息」里打开「工具批准诊断」查看 permission-request.log（含 ide、hookEvent、diagnostic 等）。")
-                }
-                .font(.callout)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var switchEffectivenessBox: some View {
-        let agentReady = agentManager.isInstalled && agentManager.isRunning && agentManager.hooksInstalled
-        let hasAnyMissing = !agentManager.isInstalled || !agentManager.isRunning || !agentManager.hooksInstalled
-        GroupBox(agentReady ? "已生效" : "未生效") {
-            VStack(alignment: .leading, spacing: 10) {
-                HStack(spacing: 8) {
-                    Image(systemName: agentReady ? "checkmark.seal.fill" : "exclamationmark.circle.fill")
-                        .foregroundStyle(agentReady ? .green : .orange)
-                    Text(agentReady
-                         ? "Agent 与 Hook 已就位时，拨杆会参与 Claude 的 PermissionRequest 与 Cursor 的 preToolUse 等批准链。"
-                         : "拨杆在 IDE 中生效需先安装 Agent 与 Hook，并把蓝牙交给 Agent；否则仅为状态显示。")
-                        .font(.callout)
-                }
-
-                if hasAnyMissing {
-                    VStack(alignment: .leading, spacing: 4) {
-                        agentChecklistRow(label: "LaunchAgent 已安装", ok: agentManager.isInstalled)
-                        agentChecklistRow(label: "Agent 已连接蓝牙", ok: agentManager.isRunning)
-                        agentChecklistRow(label: "Claude / Cursor Hook 已配置", ok: agentManager.hooksInstalled)
-                    }
-                    .padding(.leading, 4)
-
-                    HStack(spacing: 8) {
-                        if !agentManager.isInstalled {
-                            Button("安装 Agent + Hook") {
-                                agentManager.install()
-                            }
-                            .buttonStyle(.borderedProminent)
-                            .controlSize(.small)
-                        } else if !agentManager.isRunning {
-                            // 与「设备信息」相同：在 launchd 中 load + start 守护进程。
-                            // 若当前由本 App 占用蓝牙，此处也应引导先去设备信息把「蓝牙连接」切给 Agent，否则与主流程二选一相冲突（故与 DeviceInfo 同样禁用直接启动）。
-                            Button("启动 Agent") {
-                                agentManager.start()
-                            }
-                            .buttonStyle(.borderedProminent)
-                            .controlSize(.small)
-                            .disabled(agentManager.bluetoothConnectionOwner == .ahaKeyStudio)
-                            .help(
-                                agentManager.bluetoothConnectionOwner == .ahaKeyStudio
-                                ? "当前由本 App 占用蓝牙。请打开下方「设备信息…」，在「蓝牙连接」里选「由 Agent 占用」后再启 Agent；与设备信息里「启动」按钮规则一致。"
-                                : "与「设备信息」中的启动相同，由 launchd 加载并执行 ahakeyconfig-agent。"
-                            )
-                        }
-                        Button("设备信息（蓝牙 / 启停 Agent）") {
-                            switchWorkspace(.deviceInfo)
-                        }
-                        .buttonStyle(.bordered)
-                        .controlSize(.small)
-                    }
-                }
-            }
-            .padding(.top, 4)
-        }
-    }
-
-    private func agentChecklistRow(label: String, ok: Bool) -> some View {
-        HStack(spacing: 6) {
-            Image(systemName: ok ? "checkmark.circle.fill" : "xmark.circle")
-                .foregroundStyle(ok ? .green : .secondary)
-                .font(.caption)
-            Text(label)
-                .font(.caption)
-                .foregroundStyle(ok ? .primary : .secondary)
-        }
-    }
-
-
-    private var statusBar: some View {
-        HStack(spacing: 16) {
-            Label(statusBarSelectionText, systemImage: statusBarSelectionIcon)
-                .font(AhaKeyUI.Font.subhead)
-            Divider()
-                .frame(height: 14)
-            Label("蓝牙延迟 14 ms", systemImage: "rectangle.fill")
-                .font(AhaKeyUI.Font.subhead)
-                .foregroundStyle(.secondary)
-            Divider()
-                .frame(height: 14)
-            Text("固件 v\(bleManager.firmwareMainVersion).\(bleManager.firmwareSubVersion)")
-                .font(AhaKeyUI.Font.subhead)
-                .foregroundStyle(.secondary)
-            Divider()
-                .frame(height: 14)
-            Text("AI 引擎 本地离线")
-                .font(AhaKeyUI.Font.subhead)
-                .foregroundStyle(.secondary)
-            Divider()
-                .frame(height: 14)
-            Text(dirtyCount == 0 ? "配置已就绪" : "未同步改动 \(dirtyCount)")
-                .font(AhaKeyUI.Font.subhead)
-                .foregroundStyle(dirtyCount == 0 ? Color.secondary : Color.orange)
-            Spacer()
-            Button("新手引导") {
-                NotificationCenter.default.post(name: .ahaKeyDebugShowOnboardingPreview, object: nil)
-            }
-            .buttonStyle(AhaKeySecondaryButtonStyle())
-            Button("权限引导") {
-                voiceRelay.showsPermissionOnboarding = true
-            }
-            .buttonStyle(AhaKeySecondaryButtonStyle())
-        }
-        .padding(.horizontal, 18)
-        .padding(.vertical, 10)
-        .background(chromeBarBackground)
-    }
-
-    private var voiceAgentConfigurationSheet: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            HStack {
-                Text("VoiceAgent 设置")
-                    .font(.headline)
-                Spacer()
-                Button("关闭") {
-                    showsVoiceAgentConfiguration = false
-                }
-            }
-            .padding(16)
-            Divider()
-            ScrollView {
-                VStack(alignment: .leading, spacing: 20) {
-                    // 新手指引
-                    setupGuideSection
-
-                    GroupBox("AI 模型") {
-                        LLMConfigView()
-                            .padding(.top, 4)
-                            .frame(maxWidth: .infinity, alignment: .leading)
-                    }
-
-                    GroupBox("飞书 / Lark") {
-                        FeishuSetupView()
-                            .padding(.top, 4)
-                            .frame(maxWidth: .infinity, alignment: .leading)
-                    }
-                }
-                .padding(16)
-            }
-        }
-        .frame(width: 640, height: 760)
-    }
-
-    private var setupGuideSection: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack(spacing: 8) {
-                Image(systemName: "sparkles")
-                    .foregroundStyle(.purple)
-                Text("快速开始")
-                    .font(.title3.weight(.semibold))
-            }
-
-            Text("按照以下步骤完成配置，即可通过语音让 AI 助手帮你发飞书消息。")
-                .font(.callout)
-                .foregroundStyle(.secondary)
-
-            VStack(alignment: .leading, spacing: 14) {
-                guideStep(
-                    number: 1,
-                    title: "配置 AI 模型",
-                    description: "在下方「AI 模型」区域填入你的 API Key。支持 OpenAI、Claude（需代理）、Deepseek、通义等兼容接口。",
-                    isDone: llmAPIKeyConfigured
-                )
-
-                guideStep(
-                    number: 2,
-                    title: "配置飞书",
-                    description: "在飞书开放平台创建应用获取 App ID 和 Secret，填入下方「飞书」区域后扫码登录。\n需要的权限：im:message（消息）、contact:user:search（通讯录）。",
-                    link: ("打开飞书开放平台", "https://open.feishu.cn/app"),
-                    isDone: larkCLIInstalled
-                )
-
-                guideStep(
-                    number: 3,
-                    title: "添加联系人（可选）",
-                    description: "预设常用群聊或联系人可以加速消息发送。也可以不设置，直接说联系人名字，AI 会自动搜索。",
-                    isDone: false
-                )
-            }
-            .padding(12)
-            .background(
-                RoundedRectangle(cornerRadius: 8)
-                    .fill(Color(nsColor: .controlBackgroundColor))
-            )
-        }
-    }
-
-    private func guideStep(
-        number: Int,
-        title: String,
-        description: String,
-        code: String? = nil,
-        link: (String, String)? = nil,
-        isDone: Bool
-    ) -> some View {
-        HStack(alignment: .top, spacing: 10) {
-            ZStack {
-                Circle()
-                    .fill(isDone ? Color.green : Color.blue.opacity(0.15))
-                    .frame(width: 24, height: 24)
-                if isDone {
-                    Image(systemName: "checkmark")
-                        .font(.caption.weight(.bold))
-                        .foregroundStyle(.white)
-                } else {
-                    Text("\(number)")
-                        .font(.caption.weight(.bold))
-                        .foregroundStyle(Color.accentColor)
-                }
-            }
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .font(.callout.weight(.semibold))
-                Text(description)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-
-                if let code {
-                    Text(code)
-                        .font(.caption.monospaced())
-                        .textSelection(.enabled)
-                        .padding(6)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .background(RoundedRectangle(cornerRadius: 4).fill(Color.black.opacity(0.05)))
-                }
-
-                if let link {
-                    Button(link.0) {
-                        if let url = URL(string: link.1) {
-                            NSWorkspace.shared.open(url)
-                        }
-                    }
-                    .controlSize(.small)
-                }
-            }
-        }
-    }
-
-    private var larkCLIInstalled: Bool {
-        // 检查 app bundle 内置或系统安装的 lark-cli
-        var paths: [String] = []
-        if let bundlePath = Bundle.main.executableURL?
-            .deletingLastPathComponent()
-            .appendingPathComponent("lark-cli").path {
-            paths.append(bundlePath)
-        }
-        paths += [
-            "/usr/local/bin/lark-cli",
-            "/opt/homebrew/bin/lark-cli",
-            ProcessInfo.processInfo.environment["HOME"].map { "\($0)/.npm-global/bin/lark-cli" },
-        ].compactMap { $0 }
-        return paths.contains { FileManager.default.fileExists(atPath: $0) }
-    }
-
-    private var llmAPIKeyConfigured: Bool {
-        VoiceAgentRuntimeConfig.openAIAPIKey != nil
-    }
-
-    private func settingRow(title: String, value: String) -> some View {
-        HStack(alignment: .firstTextBaseline, spacing: 12) {
-            Text(title)
-                .font(.callout.weight(.semibold))
-                .frame(width: 110, alignment: .leading)
-            Text(value)
-                .font(.callout)
-                .foregroundStyle(.secondary)
-                .textSelection(.enabled)
-            Spacer(minLength: 0)
-        }
-    }
-
-    private var chromeBarBackground: Color {
-        AhaKeyUI.ColorToken.card
-    }
-
-    private var currentModeDraft: AhaKeyModeDraft {
-        studioDraft.draft(for: selectedMode)
-    }
-
-    private var currentSelectedKey: AhaKeyKeyDraft {
-        let role = selectedPart.keyRole ?? .voice
-        return currentModeDraft.key(for: role)
-    }
-
-    private var currentSwitchTitle: String {
-        bleManager.switchState == 0 ? "自动批准" : "手动批准"
-    }
-
-    private var workModeDisplayName: String {
-        switch bleManager.workMode {
-        case 0: return "Mode 0"
-        case 1: return "Mode 1"
-        case 2: return "Mode 2"
-        default: return "Mode \(bleManager.workMode)"
-        }
-    }
-
-    private var switchStateDisplayName: String {
-        bleManager.switchState == 0 ? "自动批准" : "手动批准"
-    }
-
-    private var currentOLEDPreviewImage: NSImage? {
-        guard let path = currentModeDraft.oled.localAssetPath else { return nil }
-        return NSImage(contentsOfFile: path)
-    }
-
-    private var currentOLEDAssetURL: URL? {
-        guard let path = currentModeDraft.oled.localAssetPath else { return nil }
-        return URL(fileURLWithPath: path)
-    }
-
-    private var currentLightEffect: LightEffectStyle {
-        AhaKeyLightBarDraft.hardwareEffect(for: lightBarPreview)
-    }
-
-    private var currentDevicePreviewIDEState: IDEState? {
-        currentLightEffect.previewIDEState(forSwitchState: bleManager.switchState)
-    }
-
-    private var currentLightPreviewHint: String {
-        currentLightEffect.previewHint(forSwitchState: bleManager.switchState)
-    }
-
-    private var isEditingConfiguration: Bool {
-        agentManager.bluetoothConnectionOwner == .ahaKeyStudio
-    }
-
-    private var configurationModeTitle: String {
-        if selectedMode == .mode2 {
-            return "VoiceAgent"
-        }
-        return isEditingConfiguration ? "编辑配置中" : "键盘控制中"
-    }
-
-    private var configurationModeDetail: String {
-        if selectedMode == .mode2 {
-            return "主 agent 工作台"
-        }
-        if isEditingConfiguration {
-            if bleManager.isConnected {
-                return "AhaKey Studio 正在配置键盘"
-            }
-            return bleManager.isScanning ? "AhaKey Studio 正在连接键盘" : "AhaKey Studio 等待连接键盘"
-        }
-        if agentManager.isRunning && agentManager.isAgentBLEConnected {
-            return "Agent 正在控制键盘"
-        }
-        if agentManager.isRunning {
-            return "Agent 运行中，等待键盘连接"
-        }
-        if agentManager.isInstalled {
-            return "Agent 已安装，正在准备控制"
-        }
-        return "需要安装 Agent 后才能控制键盘"
-    }
-
-    private var configurationModeButtonTitle: String {
-        if selectedMode == .mode2 {
-            return "VoiceAgent 设置"
-        }
-        if isSyncing {
-            return "同步中…"
-        }
-        if isEditingConfiguration {
-            return "返回控制"
-        }
-        return "编辑配置"
-    }
-
-    private var configurationModeButtonHelp: String {
-        if selectedMode == .mode2 {
-            return "打开 VoiceAgent 运行时配置入口；当前 API key 暂时从 Keychain 读取。"
-        }
-        if isEditingConfiguration {
-            if hasUnsyncedChanges {
-                return "将当前草稿同步到键盘，然后把蓝牙交还给 Agent。"
-            }
-            return "没有未同步改动，直接把蓝牙交还给 Agent。"
-        }
-        return "临时由 AhaKey Studio 接管蓝牙，用于改键、OLED、同步和本机灯效测试。"
-    }
-
-    private var statusBarSelectionText: String {
-        if selectedMode == .mode2 {
-            return "VoiceAgent · \(selectedMode.title)"
-        }
-        return "\(selectedPart.title) · \(selectedMode.title)"
-    }
-
-    private var statusBarSelectionIcon: String {
-        if selectedMode == .mode2 {
-            return "point.3.connected.trianglepath.dotted"
-        }
-        return selectedPart.systemImage
-    }
-
-    private var syncToKeyboardButtonTitle: String {
-        if isSyncing { return "同步中…" }
-        return hasUnsyncedChanges ? "同步到键盘" : "已同步"
-    }
-
-    private var keyConfigSyncButtonTitle: String {
-        if isSyncing { return "同步中…" }
-        if !isEditingConfiguration { return "先编辑配置" }
-        if !bleManager.isConnected { return "等待连接" }
-        return hasUnsyncedChanges ? "同步到设备" : "已同步"
-    }
-
-    private var syncToKeyboardButtonHelp: String {
-        if !bleManager.isConnected {
-            return "键盘未连接，当前只能保存本地草稿。"
-        }
-        if !hasUnsyncedChanges {
-            return "当前配置已写入键盘，无需再次同步。"
-        }
-        return "将当前 Mode、按键、OLED 和灯效草稿写入键盘。"
-    }
-
-    private var canSyncConfiguration: Bool {
-        isEditingConfiguration &&
-        hasUnsyncedChanges &&
-        bleManager.isConnected &&
-        !isSyncing &&
-        !agentManager.isAgentOperationInProgress
-    }
-
-    private var estimatedVoiceSessions: Int {
-        var count = 0
-        if nativeSpeech.isRecording { count += 1 }
-        if !nativeSpeech.lastCommittedText.isEmpty { count += 1 }
-        if !nativeSpeech.transcriptPreview.isEmpty { count += 1 }
-        return count
-    }
-
-    private var lastCommittedCharacterCountText: String {
-        let count = nativeSpeech.lastCommittedText.count
-        return count == 0 ? "—" : "\(count)"
-    }
-
-    private var voicePresetDetail: String {
-        let preset = currentSelectedKey.voicePreset ?? .custom
-        return preset.detail
-    }
-
-    private func permissionBadge(title: String, granted: Bool) -> some View {
-        HStack(spacing: 6) {
-            Circle()
-                .fill(granted ? Color.green : Color.orange)
-                .frame(width: 8, height: 8)
-            Text(title)
-                .font(.caption.weight(.semibold))
-            Text(granted ? "已开启" : "未开启")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-        }
-        .ahaKeyPill(accent: granted ? .green : .orange)
-    }
-
-    private var currentSelectedKeySanitizedDescription: String {
-        currentSelectedKey.description.sanitizedASCII(maxLength: 20)
-    }
-
-    private var selectedKeyDescriptionBinding: Binding<String> {
-        Binding(
-            get: { currentSelectedKey.description },
-            set: { newValue in
-                updateSelectedKey { key in
-                    key.description = String(newValue.prefix(20))
-                }
-            }
-        )
-    }
-
-    private var selectedKeyShortcutBinding: Binding<ShortcutBinding> {
-        Binding(
-            get: { currentSelectedKey.shortcut },
-            set: { newValue in
-                updateSelectedKey { $0.shortcut = newValue }
-            }
-        )
-    }
-
-    private var selectedVoicePresetBinding: Binding<VoicePreset> {
-        Binding(
-            get: { currentSelectedKey.voicePreset ?? .custom },
-            set: { preset in
-                guard preset.availableInV1 else { return }
-                applyVoicePreset(preset)
-            }
-        )
-    }
-
-    private var oledFramesPerSecondBinding: Binding<Int> {
-        Binding(
-            get: { currentModeDraft.oled.framesPerSecond },
-            set: { newValue in
-                updateCurrentMode { mode in
-                    mode.oled.framesPerSecond = min(20, max(5, newValue))
-                }
-            }
-        )
-    }
-
-    private var hasUnsyncedChanges: Bool {
-        dirtyCount > 0
-    }
-
-    private var dirtyCount: Int {
-        AhaKeyModeSlot.allCases.reduce(into: 0) { count, mode in
-            let current = studioDraft.draft(for: mode)
-            let baseline = lastSyncedDraft.draft(for: mode)
-            for role in AhaKeyKeyRole.allCases where current.key(for: role) != baseline.key(for: role) {
-                count += 1
-            }
-            if current.oled != baseline.oled {
-                count += 1
-            }
-        }
-    }
-
-    private func restoreCurrentModeDefaults() {
-        let restored = AhaKeyModeDraft.default(for: selectedMode)
-        var next = studioDraft
-        next.updateMode(restored)
-        studioDraft = next
-        syncStatusMessage = "\(selectedMode.title) 已恢复默认值，等待同步。"
-    }
-
-    private func clearCurrentOLED() {
-        updateCurrentMode { mode in
-            mode.oled.localAssetPath = nil
-            mode.oled.statusLine = AhaKeyOLEDDraft.default(for: selectedMode).statusLine
-        }
-    }
-
-    private func applyVoicePreset(_ preset: VoicePreset) {
-        updateSelectedKey { key in
-            key.voicePreset = preset
-            if preset != .custom {
-                key.shortcut = preset.defaultBinding
-            }
-            if key.description.isEmpty {
-                key.description = key.role.defaultDescription
-            }
-        }
-    }
-
-    // MARK: - 宏编辑
-
-    /// 按键当前处于 "宏" 还是 "快捷键" 录入模式。
-    /// 状态仅由 `macro` 是否为空推导，避免多出一个独立 flag。
-    private enum KeyBindingMode {
-        case shortcut
-        case macro
-    }
-
-    private var selectedKeyBindingModeBinding: Binding<KeyBindingMode> {
-        Binding(
-            get: { currentSelectedKey.usesMacro ? .macro : .shortcut },
-            set: { newValue in
-                switch newValue {
-                case .shortcut:
-                    updateSelectedKey { key in
-                        key.macro = []
-                    }
-                case .macro:
-                    updateSelectedKey { key in
-                        guard key.macro.isEmpty else { return }
-                        // Mode 0「No」键的 shortcut 故意为空，实际绑定是固件宏 ↓↓⏎；若仍用「空 shortcut → Enter 种子」，
-                        // 从「单键」切回「宏」时会被误植成只按 Enter，覆盖用户刚配好的三键宏。
-                        if selectedMode == .mode0, key.role == .reject {
-                            key.macro = AhaKeyModeDraft.claudeNoMacroSteps.map { step in
-                                MacroStep(action: step.action, param: step.param)
-                            }
-                            return
-                        }
-                        // 其它键：用当前 shortcut 的主键作种子（没配就用 Enter），避免空白宏列表。
-                        let seed: UInt8 = key.shortcut.keyCode == 0 ? HIDUsage.enter : key.shortcut.keyCode
-                        key.macro = [
-                            MacroStep(action: .downKey, param: seed),
-                            MacroStep(action: .upKey, param: seed),
-                        ]
-                    }
-                }
-            }
-        )
-    }
-
-    private func appendMacroStep() {
-        updateSelectedKey { key in
-            // 默认追加 "按下 Enter"——多数用户添加步骤都是想按键，延时/松开可以再切。
-            key.macro.append(MacroStep(action: .downKey, param: HIDUsage.enter))
-        }
-    }
-
-    private func removeMacroStep(at index: Int) {
-        updateSelectedKey { key in
-            guard key.macro.indices.contains(index) else { return }
-            key.macro.remove(at: index)
-        }
-    }
-
-    private func moveMacroStep(from index: Int, by offset: Int) {
-        updateSelectedKey { key in
-            let target = index + offset
-            guard key.macro.indices.contains(index), key.macro.indices.contains(target) else { return }
-            key.macro.swapAt(index, target)
-        }
-    }
-
-    private func updateMacroStep(id: UUID, transform: (inout MacroStep) -> Void) {
-        updateSelectedKey { key in
-            guard let idx = key.macro.firstIndex(where: { $0.id == id }) else { return }
-            transform(&key.macro[idx])
-        }
-    }
-
-    private func updateSelectedKey(_ transform: (inout AhaKeyKeyDraft) -> Void) {
-        guard let role = selectedPart.keyRole else { return }
-        updateCurrentMode { mode in
-            var key = mode.key(for: role)
-            transform(&key)
-            mode.updateKey(key)
-        }
-    }
-
-    private func updateCurrentMode(_ transform: (inout AhaKeyModeDraft) -> Void) {
-        updateMode(selectedMode, transform)
-    }
-
-    private func updateMode(_ modeSlot: AhaKeyModeSlot, _ transform: (inout AhaKeyModeDraft) -> Void) {
-        var next = studioDraft
-        var mode = next.draft(for: modeSlot)
-        transform(&mode)
-        next.updateMode(mode)
-        studioDraft = next
-    }
-
-    private func partIsDirty(_ part: AhaKeyStudioPart) -> Bool {
-        let current = studioDraft.draft(for: selectedMode)
-        let baseline = lastSyncedDraft.draft(for: selectedMode)
-        switch part {
-        case .key1, .key2, .key3, .key4:
-            guard let role = part.keyRole else { return false }
-            return current.key(for: role) != baseline.key(for: role)
-        case .oledDisplay:
-            return current.oled != baseline.oled
-        case .lightBar, .toggleSwitch:
-            return false
-        }
-    }
-
-    private func dirtyPartsForCurrentMode() -> Set<AhaKeyStudioPart> {
-        Set(AhaKeyStudioPart.allCases.filter(partIsDirty(_:)))
-    }
-
-    private func selectOLEDGIF() {
-        let panel = NSOpenPanel()
-        panel.allowedContentTypes = [UTType(filenameExtension: "gif")!]
-        panel.allowsMultipleSelection = false
-        if panel.runModal() == .OK, let url = panel.url {
-            do {
-                try OLEDFrameEncoder.validateGIFSourceFileSize(at: url)
-            } catch {
-                let msg = (error as? LocalizedError)?.errorDescription ?? "GIF 文件过大。"
-                syncStatusMessage = msg
-                updateCurrentMode { mode in
-                    mode.oled.statusLine = msg
-                }
-                return
-            }
-            let frameCount = OLEDFrameEncoder.frameCount(at: url)
-            updateCurrentMode { mode in
-                mode.oled.localAssetPath = url.path
-                mode.oled.statusLine = "已选 \(max(frameCount, 1)) 帧 GIF 预览；切换模式时会先显示描述，再回到当前模式动图。"
-            }
-            syncStatusMessage = "已更新 \(selectedMode.title) 的 OLED 预览，连接后可直接上传到设备。"
-        }
-    }
-
-    private func handleConfigurationModeButton() {
-        if selectedMode == .mode2 {
-            openVoiceAgentConfiguration()
-            return
-        }
-        if isEditingConfiguration {
-            finishEditingConfiguration()
-        } else {
-            enterEditingConfiguration()
-        }
-    }
-
-    private func openVoiceAgentConfiguration() {
-        showsVoiceAgentConfiguration = true
-        syncStatusMessage = "VoiceAgent 配置入口已预留，当前运行时读取 Keychain。"
-    }
-
-    private func enterEditingConfiguration() {
-        agentManager.setBluetoothConnectionOwner(.ahaKeyStudio, bleManager: bleManager)
-        syncStatusMessage = "已进入编辑配置，AhaKey Studio 将临时接管蓝牙。"
-    }
-
-    private func finishEditingConfiguration() {
-        guard hasUnsyncedChanges else {
-            returnToKeyboardControl()
-            return
-        }
-
-        guard bleManager.isConnected else {
-            syncStatusMessage = "键盘尚未连接，无法同步未保存改动。请等待连接成功后再返回控制。"
-            bleManager.userInitiatedConnect()
-            return
-        }
-
-        syncAllModesToDevice(returnToKeyboardControlWhenDone: true)
-    }
-
-    private func returnToKeyboardControl() {
-        agentManager.setBluetoothConnectionOwner(.agentDaemon, bleManager: bleManager)
-        syncStatusMessage = "已返回键盘控制，Agent 将接管蓝牙。"
-    }
-
-    private func syncAllModesToDevice(returnToKeyboardControlWhenDone: Bool = false) {
-        guard bleManager.isConnected else {
-            syncStatusMessage = "设备未连接，当前只保存本地草稿。"
-            return
-        }
-
-        applyCursorRejectMacroSelfHealIfNeeded()
-        var commands = commandsForModes(AhaKeyModeSlot.allCases)
-        commands.append((data: AhaKeyCommand.saveConfig(), label: "保存全部配置到设备"))
-
-        let total = commands.count
-        isSyncing = true
-        syncStatusMessage = "正在写入设备（约 \(total) 条，全部发完后再保存/交还 Agent）…"
-        let returnAgent = returnToKeyboardControlWhenDone
-        bleManager.writeCommandsSequentially(commands) {
-            Task { @MainActor in
-                // 队列与 50ms 间隔已保证顺序；略等再交还蓝牙，避免固件尚未处理完最后帧。
-                try? await Task.sleep(for: .milliseconds(250))
-                self.lastSyncedDraft = self.studioDraft
-                self.lastSyncDate = Date()
-                self.isSyncing = false
-                self.syncStatusMessage = "已全部写入设备并保存。"
-                if returnAgent {
-                    self.returnToKeyboardControl()
-                }
-            }
-        }
-    }
-
-    private func resendCurrentModeToDevice() {
-        guard bleManager.isConnected else {
-            syncStatusMessage = "设备未连接，当前只保存本地草稿。"
-            return
-        }
-
-        applyCursorRejectMacroSelfHealIfNeeded()
-        var commands = commandsForModes([selectedMode])
-        commands.append((data: AhaKeyCommand.saveConfig(), label: "保存 \(selectedMode.title) 当前配置"))
-
-        isSyncing = true
-        syncStatusMessage = "正在写入 \(selectedMode.title)…"
-        bleManager.writeCommandsSequentially(commands) {
-            Task { @MainActor in
-                try? await Task.sleep(for: .milliseconds(150))
-                self.lastSyncDate = Date()
-                self.isSyncing = false
-                self.syncStatusMessage = "已重新发送 \(self.selectedMode.title) 当前模式。"
-            }
-        }
-    }
-
-    /// Cursor 档「取消键」若仍为默认 ⌫ 却残留宏，同步会走 0x74 而非单键。清掉误残留宏并与迁移逻辑一致。
-    private func applyCursorRejectMacroSelfHealIfNeeded() {
-        var next = studioDraft
-        var m1 = next.draft(for: .mode1)
-        var reject = m1.key(for: .reject)
-        let defaultR = AhaKeyModeDraft.default(for: .mode1).key(for: .reject)
-        guard !reject.macro.isEmpty, reject.shortcut == defaultR.shortcut else { return }
-        reject.macro = []
-        m1.updateKey(reject)
-        next.updateMode(m1)
-        studioDraft = next
-    }
-
-    private func commandsForModes(_ modes: [AhaKeyModeSlot]) -> [(data: Data, label: String)] {
-        var commands: [(data: Data, label: String)] = []
-
-        for mode in modes {
-            let draft = studioDraft.draft(for: mode)
-            for role in AhaKeyKeyRole.allCases {
-                let key = draft.key(for: role)
-                let keyIndex = UInt8(role.rawValue)
-                let modeByte = UInt8(mode.rawValue)
-
-                if key.usesMacro {
-                    // 固件对 0x73 快捷键、0x74 宏是分层存储的；从「快捷键」改「宏」时须先清掉旧快捷键，否则会残留。
-                    commands.append((
-                        data: AhaKeyCommand.setKeyMapping(
-                            mode: modeByte,
-                            keyIndex: keyIndex,
-                            hidCodes: []
-                        ),
-                        label: "清除 \(mode.title) \(key.title) 快捷键层（将写入宏）"
-                    ))
-                    commands.append((
-                        data: AhaKeyCommand.setKeyMacro(
-                            mode: modeByte,
-                            keyIndex: keyIndex,
-                            macroData: key.macro.flattenedBytes
-                        ),
-                        label: "写入 \(mode.title) \(key.title) 宏: \(key.macro.displaySummary)"
-                    ))
-                } else {
-                    // 从「宏」改「快捷键 / 无键」时须先发空 0x74，否则设备可能仍走旧宏（Cursor/其它 mode 上表现为改键不生效）。
-                    commands.append((
-                        data: AhaKeyCommand.setKeyMacro(
-                            mode: modeByte,
-                            keyIndex: keyIndex,
-                            macroData: []
-                        ),
-                        label: "清除 \(mode.title) \(key.title) 宏层（将写入快捷键）"
-                    ))
-                    if !key.shortcut.hidCodes.isEmpty {
-                        commands.append((
-                            data: AhaKeyCommand.setKeyMapping(
-                                mode: modeByte,
-                                keyIndex: keyIndex,
-                                hidCodes: key.shortcut.hidCodes
-                            ),
-                            label: "写入 \(mode.title) \(key.title) 快捷键: \(key.shortcut.displayLabel)"
-                        ))
-                    } else {
-                        commands.append((
-                            data: AhaKeyCommand.setKeyMapping(
-                                mode: modeByte,
-                                keyIndex: keyIndex,
-                                hidCodes: []
-                            ),
-                            label: "清除 \(mode.title) \(key.title) 快捷键"
-                        ))
-                    }
-                }
-
-                let sanitizedDescription = key.description.sanitizedASCII(maxLength: 20)
-                commands.append((
-                    data: AhaKeyCommand.setKeyDescription(
-                        mode: UInt8(mode.rawValue),
-                        keyIndex: keyIndex,
-                        text: key.description
-                    ),
-                    label: "写入 \(mode.title) \(key.title) 描述: \(sanitizedDescription.isEmpty ? "空白" : sanitizedDescription)"
-                ))
-            }
-        }
-
-        return commands
-    }
-
-    private func uploadCurrentOLEDToDevice() {
-        guard bleManager.isConnected else {
-            syncStatusMessage = "设备未连接，先连上键盘再上传 OLED 动图。"
-            return
-        }
-        guard let assetPath = currentModeDraft.oled.localAssetPath else {
-            syncStatusMessage = "先为 \(selectedMode.title) 选择一个 GIF，再上传到设备。"
-            return
-        }
-
-        let targetMode = selectedMode
-        let targetFPS = currentModeDraft.oled.framesPerSecond
-        let assetURL = URL(fileURLWithPath: assetPath)
-
-        updateMode(targetMode) { mode in
-            mode.oled.statusLine = "正在上传动图到 \(targetMode.title)…"
-        }
-        syncStatusMessage = "开始上传 \(targetMode.title) 的 OLED 动图。"
-
-        Task { @MainActor in
-            do {
-                let frames = try OLEDFrameEncoder.frames(fromGIFAt: assetURL)
-                let startIndex = try await resolveOLEDUploadStartIndex(for: targetMode, frameCount: frames.count)
-                try await bleManager.uploadOLEDFrames(
-                    frames,
-                    fps: targetFPS,
-                    mode: UInt8(targetMode.rawValue),
-                    startIndex: UInt16(startIndex)
-                )
-                updateMode(targetMode) { mode in
-                    mode.oled.statusLine = "已上传 \(frames.count) 帧到设备，槽位起点 \(startIndex)；切换模式时会先显示描述，再回到当前模式动图。"
-                }
-                syncStatusMessage = "\(targetMode.title) OLED 动图已上传完成。"
-            } catch {
-                updateMode(targetMode) { mode in
-                    mode.oled.statusLine = "上传失败：\(error.localizedDescription)"
-                }
-                syncStatusMessage = "\(targetMode.title) OLED 上传失败：\(error.localizedDescription)"
-            }
-        }
-    }
-
-    private func resolveOLEDUploadStartIndex(for targetMode: AhaKeyModeSlot, frameCount: Int) async throws -> Int {
-        var states: [AhaKeyPictureState] = []
-        for mode in AhaKeyModeSlot.allCases {
-            states.append(try await bleManager.readPictureState(mode: UInt8(mode.rawValue)))
-        }
-
-        let maxCapacity = states.first?.allModeMaxPic ?? AhaKeyCommand.oledMaxFrames
-        guard frameCount <= maxCapacity else {
-            throw OLEDUploadError.noAvailablePictureSlot(needed: frameCount, max: maxCapacity)
-        }
-
-        let currentState = states.first(where: { $0.mode == targetMode.rawValue })
-        let occupiedRegions = states
-            .filter { $0.mode != targetMode.rawValue && $0.picLength > 0 }
-            .map { (start: $0.startIndex, end: $0.startIndex + $0.picLength) }
-            .sorted { $0.start < $1.start }
-
-        if let currentState,
-           currentState.picLength > 0,
-           canPlacePictureRange(
-               start: currentState.startIndex,
-               count: frameCount,
-               occupiedRegions: occupiedRegions,
-               maxCapacity: maxCapacity
-           )
-        {
-            return currentState.startIndex
-        }
-
-        if let freeStart = findFreePictureSpace(
-            occupiedRegions: occupiedRegions,
-            neededCount: frameCount,
-            maxCapacity: maxCapacity
-        ) {
-            return freeStart
-        }
-
-        throw OLEDUploadError.noAvailablePictureSlot(needed: frameCount, max: maxCapacity)
-    }
-
-    private func canPlacePictureRange(
-        start: Int,
-        count: Int,
-        occupiedRegions: [(start: Int, end: Int)],
-        maxCapacity: Int
-    ) -> Bool {
-        let end = start + count
-        guard start >= 0, end <= maxCapacity else { return false }
-        return occupiedRegions.allSatisfy { region in
-            end <= region.start || start >= region.end
-        }
-    }
-
-    private func findFreePictureSpace(
-        occupiedRegions: [(start: Int, end: Int)],
-        neededCount: Int,
-        maxCapacity: Int
-    ) -> Int? {
-        guard !occupiedRegions.isEmpty else { return 0 }
-
-        if occupiedRegions[0].start >= neededCount {
-            return 0
-        }
-
-        for index in 0 ..< (occupiedRegions.count - 1) {
-            let gapStart = occupiedRegions[index].end
-            let gapEnd = occupiedRegions[index + 1].start
-            if gapEnd - gapStart >= neededCount {
-                return gapStart
-            }
-        }
-
-        let lastEnd = occupiedRegions.last?.end ?? 0
-        if lastEnd + neededCount <= maxCapacity {
-            return lastEnd
-        }
-
-        return nil
-    }
-
-    private func previewCurrentLightEffectOnDevice() {
-        guard let ideState = currentDevicePreviewIDEState else {
-            syncStatusMessage = currentLightPreviewHint
-            return
-        }
-        bleManager.updateIDEState(ideState)
-        if bleManager.workMode == 0 {
-            syncStatusMessage = "已把 \(currentLightEffect.title) 的可达预览发送到设备（Mode 0）。"
-        } else {
-            syncStatusMessage = "已发送 \(ideState.label) 到设备。注意：键盘当前在 Mode \(bleManager.workMode)，出厂固件可能没有在此档位映射此 state，把拨杆切到 Mode 0 可看到完整效果。"
-        }
-    }
-
-    private func copyBLELog() {
-        let text = bleManager.commLog.map { "[\($0.formattedTime)] \($0.message)" }.joined(separator: "\n")
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(text, forType: .string)
-    }
-
-    private func infoPill(title: String, subtitle: String, accent: Color) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            Text(title)
-                .font(AhaKeyUI.Font.caption)
-                .foregroundStyle(.secondary)
-            HStack(spacing: 6) {
-                Circle()
-                    .fill(accent)
-                    .frame(width: 8, height: 8)
-                Text(subtitle)
-                    .font(AhaKeyUI.Font.subhead.weight(.semibold))
-            }
-        }
-        .ahaKeyPill(accent: accent)
-    }
-
-    private func manualCallout(title: String, detail: String) -> some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text(title)
-                .font(.caption.weight(.semibold))
-                .foregroundStyle(.secondary)
-            Text(detail)
-                .font(.callout)
-                .lineLimit(2)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .ahaKeyRemarkPanel()
-    }
-
-    private static let timeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "HH:mm:ss"
-        return formatter
-    }()
-
-    private func openNativeSpeechPrivacySettings() {
-        openStudioNativeSpeechPrivacySettingsURL()
-    }
-}
-
-private struct VoicePermissionOnboardingSheet: View {
-    @ObservedObject var voiceRelay: VoiceRelayService
-    @ObservedObject var nativeSpeech: NativeSpeechTranscriptionService
-    @Environment(\.dismiss) private var dismiss
-
-    @State private var fixInProgress = false
-    @State private var fixAlertTitle = ""
-    @State private var fixAlertMessage = ""
-    @State private var fixAlertIsSuccess = false
-    @State private var showFixAlert = false
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 18) {
-            Text("开启语音快捷键权限")
-                .font(.system(size: 24, weight: .semibold))
-
-            Text("为了让 AhaKey 的语音键在后台直接接管语音，第一次使用时需要给 AhaKey Studio 打开系统权限。macOS 原生语音还会额外用到苹果自己的麦克风和语音转写能力。")
-                .font(.callout)
-                .foregroundStyle(.secondary)
-
-            VStack(alignment: .leading, spacing: 10) {
-                permissionRow(title: "输入监控", granted: voiceRelay.inputMonitoringGranted, detail: "允许 AhaKey Studio 在后台监听实体语音键。")
-                permissionRow(title: "辅助功能", granted: voiceRelay.accessibilityGranted, detail: "允许 AhaKey Studio 把语音键转换成苹果原生转写或 Fn/Globe。")
-                permissionRow(title: "麦克风", granted: nativeSpeech.microphoneGranted, detail: "允许 AhaKey Studio 使用苹果原生语音采集。")
-                permissionRow(title: "语音转写", granted: nativeSpeech.speechRecognitionGranted, detail: "允许 AhaKey Studio 使用苹果原生语音识别。")
-            }
-
-            Text("操作建议：先点「现在申请权限」——macOS 上输入监控/辅助功能常常不再弹系统对话框，约半秒后会自动打开「隐私与安全性」，请在列表中勾选 AhaKey Studio；麦克风和语音在之前就拒绝过的话也不会再弹窗，需在设置里手动打开。若你已在系统设置里改好，可点「我已完成，重新检查」。")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-            Text("若系统里已勾选允许，本应用仍显示未开启：请完全退出 AhaKey Studio 并再启动一次。输入监控、辅助功能等常按进程生效，只点「重新检查」或从后台切回，有时读到的仍是旧状态，重启后即可与系统设置一致。")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-            Text("外发 / DMG / Xcode：默认正式包在系统「隐私与安全性」里显示为「AhaKey Studio」；用 Xcode 以 Debug 运行本工程时显示为「AhaKey Studio（调试）」，请按名称分别授权。路径或签名不同也会被系统当成另一款 App。")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(voiceRelay.lastPermissionCheckSummary)
-                Text(nativeSpeech.lastPermissionCheckSummary)
-            }
-            .font(.caption)
-            .foregroundStyle(.secondary)
-
-            HStack(spacing: 12) {
-                Button("现在申请权限") {
-                    requestStudioPermissionsThenOpenPrivacySettingsIfNeeded(
-                        voiceRelay: voiceRelay,
-                        nativeSpeech: nativeSpeech
-                    )
-                }
-                .buttonStyle(.borderedProminent)
-
-                Button("我已完成，重新检查") {
-                    voiceRelay.refreshPermissions(deferredTCCRequery: true)
-                    nativeSpeech.refreshPermissions(deferredTCCRequery: true)
-                }
-                .buttonStyle(.bordered)
-
-                RestartToApplyPermissionsButton(title: "退出并重新打开")
-
-                if !nativeSpeech.microphoneGranted || !nativeSpeech.speechRecognitionGranted {
-                    Button("打开系统设置") {
-                        openStudioNativeSpeechPrivacySettingsURL()
-                    }
-                    .buttonStyle(.bordered)
-                }
-
-                if DebugSigningFixer.isAvailable {
-                    Button(fixInProgress ? "重置中…" : "⚙️ 重置开发环境签名（通常不需要）") {
-                        runDebugSigningFix()
-                    }
-                    .buttonStyle(.bordered)
-                    .tint(.purple)
-                    .disabled(fixInProgress)
-                    .help("仅在异常情况下使用：证书过期 / 换 Mac / Team ID 变化 / 钥匙串损坏导致权限失效时，点一下会重新签名 app 并重置 TCC 授权。正式发行版（无源码目录）看不到此按钮。")
-                }
-
-                Spacer()
-
-                Button("稍后再说") {
-                    voiceRelay.dismissPermissionOnboarding()
-                    dismiss()
-                }
-                .buttonStyle(.borderless)
-            }
-
-            if voiceRelay.inputMonitoringGranted && voiceRelay.accessibilityGranted {
-                Text("基础权限已经齐了。关闭这个弹窗后，AhaKey Studio 会继续在后台监听语音键；如果你使用 macOS 原生语音，麦克风和语音转写也建议一起打开。")
-                    .font(.caption)
-                    .foregroundStyle(.green)
-            }
-        }
-        .padding(24)
-        .frame(width: 560)
-        .onChange(of: voiceRelay.inputMonitoringGranted) { _, _ in
-            closeIfReady()
-        }
-        .onChange(of: voiceRelay.accessibilityGranted) { _, _ in
-            closeIfReady()
-        }
-        .alert(fixAlertTitle, isPresented: $showFixAlert) {
-            if fixAlertIsSuccess {
-                Button("立即退出 App") { NSApp.terminate(nil) }
-                Button("稍后再退", role: .cancel) {}
-            } else {
-                Button("好", role: .cancel) {}
-            }
-        } message: {
-            Text(fixAlertMessage)
-        }
-    }
-
-    private func runDebugSigningFix() {
-        fixInProgress = true
-        DebugSigningFixer.run { result in
-            fixInProgress = false
-            fixAlertIsSuccess = result.success
-            fixAlertTitle = result.success ? "修复完成" : "修复失败"
-            fixAlertMessage = result.output
-            showFixAlert = true
-        }
-    }
-
-    private func closeIfReady() {
-        guard voiceRelay.inputMonitoringGranted && voiceRelay.accessibilityGranted else { return }
-        voiceRelay.dismissPermissionOnboarding()
-        dismiss()
-    }
-
-    private func permissionRow(title: String, granted: Bool, detail: String) -> some View {
-        HStack(alignment: .top, spacing: 12) {
-            Circle()
-                .fill(granted ? Color.green : Color.orange)
-                .frame(width: 10, height: 10)
-                .padding(.top, 4)
-
-            VStack(alignment: .leading, spacing: 4) {
-                HStack(spacing: 8) {
-                    Text(title)
-                        .font(.headline)
-                    Text(granted ? "已开启" : "未开启")
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-                Text(detail)
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .padding(12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(
-            RoundedRectangle(cornerRadius: 14)
-                .fill(Color(nsColor: .controlBackgroundColor))
-        )
-    }
-}
-
-private struct VoicePresetPicker: View {
-    let selectedPreset: VoicePreset
-    let onSelect: (VoicePreset) -> Void
-
-    private let columns = [
-        GridItem(.flexible(), spacing: 10),
-        GridItem(.flexible(), spacing: 10),
-    ]
-
-    var body: some View {
-        LazyVGrid(columns: columns, spacing: 10) {
-            ForEach(VoicePreset.allCases) { preset in
-                Button {
-                    if preset.availableInV1 {
-                        onSelect(preset)
-                    }
-                } label: {
-                    VStack(alignment: .leading, spacing: 6) {
-                        HStack {
-                            Text(preset.title)
-                                .font(.callout.weight(.semibold))
-                            Spacer()
-                            if !preset.availableInV1 {
-                                Text("开发中")
-                                    .font(.caption2)
-                                    .foregroundStyle(.secondary)
-                            }
-                        }
-                        Text(preset.detail)
-                            .font(.caption)
-                            .foregroundStyle(.secondary)
-                            .lineLimit(3)
-                    }
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(12)
-                    .background(
-                        RoundedRectangle(cornerRadius: 12)
-                            .fill(cardFill(for: preset))
-                    )
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 12)
-                            .stroke(cardStroke(for: preset), lineWidth: preset == selectedPreset ? 1.5 : 1)
-                    )
-                }
-                .buttonStyle(.plain)
-                .disabled(!preset.availableInV1)
-            }
-        }
-    }
-
-    private func cardFill(for preset: VoicePreset) -> Color {
-        if preset == selectedPreset {
-            return Color.accentColor.opacity(0.16)
-        }
-        if !preset.availableInV1 {
-            return Color(nsColor: .controlBackgroundColor).opacity(0.65)
-        }
-        return Color(nsColor: .controlBackgroundColor)
-    }
-
-    private func cardStroke(for preset: VoicePreset) -> Color {
-        if preset == selectedPreset {
-            return .accentColor
-        }
-        return Color.black.opacity(0.08)
-    }
-}
-
-private struct ShortcutBindingEditor: View {
-    @Binding var shortcut: ShortcutBinding
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack(spacing: 8) {
-                ForEach(ShortcutModifier.allCases) { modifier in
-                    Toggle(isOn: modifierBinding(modifier)) {
-                        Text(modifier.symbol)
-                            .font(.system(.headline, design: .rounded))
-                    }
-                    .toggleStyle(.button)
-                    .help(modifier.title)
-                }
-                if !shortcut.modifiers.isEmpty {
-                    Button("清除修饰键") {
-                        var next = shortcut
-                        next.modifiers = []
-                        shortcut = next
-                    }
-                    .buttonStyle(.bordered)
-                    .controlSize(.small)
-                }
-            }
-
-            Picker("主键", selection: primaryKeyBinding) {
-                Text("未设置").tag(UInt8(0))
-                ForEach(HIDUsage.allOptions, id: \.code) { option in
-                    Text(option.name).tag(option.code)
-                }
-            }
-            .pickerStyle(.menu)
-
-            if !shortcut.modifiers.isEmpty {
-                Text("当前为组合键（\(shortcut.displayLabel)）。若你只想发单键 Enter，勿打开 ⌘/⌃ 等，或点「清除修饰键」后再选 Enter。")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-                    .fixedSize(horizontal: false, vertical: true)
-            }
-        }
-    }
-
-    private func modifierBinding(_ modifier: ShortcutModifier) -> Binding<Bool> {
-        Binding(
-            get: { shortcut.modifiers.contains(modifier) },
-            set: { on in
-                var next = shortcut
-                next.setModifier(modifier, enabled: on)
-                shortcut = next
-            }
-        )
-    }
-
-    private var primaryKeyBinding: Binding<UInt8> {
-        Binding(
-            get: { shortcut.keyCode },
-            set: { newCode in
-                var next = shortcut
-                next.keyCode = newCode
-                shortcut = next
-            }
-        )
-    }
-}
-
-private struct AhaKeyKeyboardCanvasView: View {
-    let modeDraft: AhaKeyModeDraft
-    let selectedPart: AhaKeyStudioPart
-    let lightBarPreview: LightBarPreviewState
-    let switchTitle: String
-    let dirtyParts: Set<AhaKeyStudioPart>
-    let onSelect: (AhaKeyStudioPart) -> Void
-
-    private let baseWidth: CGFloat = 109
-    private let baseHeight: CGFloat = 54
-
-    var body: some View {
-        GeometryReader { proxy in
-            let drawingWidth = min(proxy.size.width, proxy.size.height * (baseWidth / baseHeight))
-            let drawingHeight = drawingWidth * (baseHeight / baseWidth)
-
-            ZStack {
-                keyboardFrame(width: drawingWidth, height: drawingHeight)
-            }
-            .frame(width: drawingWidth, height: drawingHeight)
-            .position(x: proxy.size.width / 2, y: proxy.size.height / 2)
-        }
-    }
-
-    @ViewBuilder
-    private func keyboardFrame(width: CGFloat, height: CGFloat) -> some View {
-        ZStack(alignment: .topLeading) {
-            RoundedRectangle(cornerRadius: 24, style: .continuous)
-                .fill(
-                    LinearGradient(
-                        colors: [Color.white.opacity(0.95), Color(red: 0.92, green: 0.95, blue: 0.98)],
-                        startPoint: .topLeading,
-                        endPoint: .bottomTrailing
-                    )
-                )
-                .overlay(
-                    RoundedRectangle(cornerRadius: 24, style: .continuous)
-                        .stroke(Color.black.opacity(0.1), lineWidth: 1.2)
-                )
-                .shadow(color: .black.opacity(0.08), radius: 18, y: 14)
-
-            RoundedRectangle(cornerRadius: 18, style: .continuous)
-                .stroke(Color.black.opacity(0.06), lineWidth: 1)
-                .padding(12)
-
-            VStack {
-                Spacer()
-            }
-
-            ForEach(Array([(8.0, 8.0), (101.0, 8.0), (8.0, 46.0), (101.0, 46.0)].enumerated()), id: \.offset) { _, point in
-                Circle()
-                    .stroke(Color.black.opacity(0.14), lineWidth: 1.2)
-                    .background(Circle().fill(Color.white.opacity(0.4)))
-                    .frame(width: scaled(4.8, in: width), height: scaled(4.8, in: width))
-                    .position(position(point.0, point.1, width: width, height: height))
-            }
-
-            RoundedRectangle(cornerRadius: 10, style: .continuous)
-                .stroke(Color.black.opacity(0.12), lineWidth: 1)
-                .frame(width: scaled(4.2, in: width), height: scaled(12, in: width))
-                .position(position(3.8, 28, width: width, height: height))
-
-            RoundedRectangle(cornerRadius: 18, style: .continuous)
-                .fill(Color.black.opacity(0.04))
-                .frame(width: scaled(70, in: width), height: scaled(24, in: width))
-                .position(position(43.8, 37.3, width: width, height: height))
-
-            ledBarButton(width: width, height: height)
-            oledButton(width: width, height: height)
-            keyButton(for: .voice, width: width, height: height)
-            keyButton(for: .approve, width: width, height: height)
-            keyButton(for: .reject, width: width, height: height)
-            keyButton(for: .submit, width: width, height: height)
-            modeSwitchKey(width: width, height: height)
-            switchButton(width: width, height: height)
-        }
-    }
-
-    private func ledBarButton(width: CGFloat, height: CGFloat) -> some View {
-        let part = AhaKeyStudioPart.lightBar
-        let rect = frame(12.3, 5.0, 55.6, 9.8, width: width, height: height)
-        return Button {
-            onSelect(part)
-        } label: {
-            VStack(spacing: rect.height * 0.12) {
-                Text("灯条")
-                    .font(.system(size: max(rect.height * 0.18, 10), weight: .semibold))
-                    .foregroundStyle(Color.black.opacity(0.72))
-                    .frame(maxWidth: .infinity, alignment: .center)
-
-                ZStack(alignment: .leading) {
-                    RoundedRectangle(cornerRadius: 12, style: .continuous)
-                        .fill(Color.black.opacity(0.12))
-                    HStack(spacing: rect.width * 0.085) {
-                        ForEach(0..<4, id: \.self) { index in
-                            Capsule()
-                                .fill(ledColor(for: index))
-                                .frame(width: rect.width * 0.12, height: rect.height * 0.22)
-                        }
-                    }
-                    .padding(.horizontal, rect.width * 0.08)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                }
-                .frame(height: rect.height * 0.48)
-            }
-            .frame(width: rect.width, height: rect.height)
-            .modifier(HotspotChrome(part: part, selectedPart: selectedPart, dirtyParts: dirtyParts))
-        }
-        .buttonStyle(.plain)
-        .position(x: rect.midX, y: rect.midY)
-    }
-
-    private func oledButton(width: CGFloat, height: CGFloat) -> some View {
-        let part = AhaKeyStudioPart.oledDisplay
-        let rect = frame(71.2, 7.7, 24.2, 13.4, width: width, height: height)
-        return Button {
-            onSelect(part)
-        } label: {
-            ZStack {
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .fill(Color.black.opacity(0.92))
-                LinearGradient(
-                    colors: [Color.black.opacity(0.2), Color.white.opacity(0.05)],
-                    startPoint: .topLeading,
-                    endPoint: .bottomTrailing
-                )
-                .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-                VStack(alignment: .center, spacing: 4) {
-                    if modeDraft.oled.localAssetPath == nil {
-                        if modeDraft.mode == .mode0 {
-                            HStack(spacing: 6) {
-                                Image(systemName: "cloud.fill")
-                                    .font(.system(size: rect.height * 0.24, weight: .semibold))
-                                    .foregroundStyle(Color.orange.opacity(0.92))
-                                Text("Mode 0")
-                                    .font(.system(size: rect.height * 0.18, weight: .semibold))
-                                    .foregroundStyle(.white.opacity(0.85))
-                            }
-                            Text("默认动图")
-                                .font(.system(size: rect.height * 0.18))
-                                .foregroundStyle(.white.opacity(0.55))
-                        } else {
-                            HStack(spacing: 6) {
-                                Image(systemName: "sparkles.rectangle.stack")
-                                    .font(.system(size: rect.height * 0.22, weight: .semibold))
-                                    .foregroundStyle(.white.opacity(0.78))
-                                Text("未上传")
-                                    .font(.system(size: rect.height * 0.18, weight: .semibold))
-                                    .foregroundStyle(.white.opacity(0.85))
-                            }
-                            Text("等待自定义")
-                                .font(.system(size: rect.height * 0.18))
-                                .foregroundStyle(.white.opacity(0.55))
-                        }
-                    } else {
-                        HStack(spacing: 6) {
-                            Image(systemName: "photo")
-                                .font(.system(size: rect.height * 0.22, weight: .semibold))
-                                .foregroundStyle(.white.opacity(0.78))
-                            Text("已上传")
-                                .font(.system(size: rect.height * 0.2, weight: .semibold))
-                                .foregroundStyle(.white.opacity(0.85))
-                        }
-                        Text("预览动画中")
-                            .font(.system(size: rect.height * 0.18))
-                            .foregroundStyle(.white.opacity(0.55))
-                    }
-                }
-                .padding(rect.width * 0.08)
-                .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .center)
-                .multilineTextAlignment(.center)
-            }
-            .frame(width: rect.width, height: rect.height)
-            .modifier(HotspotChrome(part: part, selectedPart: selectedPart, dirtyParts: dirtyParts))
-        }
-        .buttonStyle(.plain)
-        .position(x: rect.midX, y: rect.midY)
-    }
-
-    private func keyButton(for role: AhaKeyKeyRole, width: CGFloat, height: CGFloat) -> some View {
-        let part = role.part
-        let keyDraft = modeDraft.key(for: role)
-        let specs: (x: CGFloat, y: CGFloat, w: CGFloat, h: CGFloat)
-        switch role {
-        case .voice:
-            specs = (10.2, 29.2, 16.2, 16.8)
-        case .approve:
-            specs = (27.2, 29.2, 16.2, 16.8)
-        case .reject:
-            specs = (44.2, 29.2, 16.2, 16.8)
-        case .submit:
-            specs = (61.2, 29.2, 16.2, 16.8)
-        }
-        let rect = frame(specs.x, specs.y, specs.w, specs.h, width: width, height: height)
-        return Button {
-            onSelect(part)
-        } label: {
-            VStack(spacing: rect.height * 0.07) {
-                ZStack {
-                    RoundedRectangle(cornerRadius: rect.width * 0.18, style: .continuous)
-                        .fill(
-                            LinearGradient(
-                                colors: [Color.white, Color(red: 0.95, green: 0.96, blue: 0.98)],
-                                startPoint: .top,
-                                endPoint: .bottom
-                            )
-                        )
-                        .shadow(color: .black.opacity(0.12), radius: 10, y: 4)
-
-                    Image(systemName: role.systemImage)
-                        .font(.system(size: rect.height * 0.24, weight: .regular))
-                        .foregroundStyle(Color.black.opacity(0.88))
-                }
-                .frame(width: rect.width * 0.8, height: rect.height * 0.76)
-
-                Text(keyDraft.description.isEmpty ? keyDraft.displaySummary : keyDraft.description)
-                    .font(.system(size: rect.height * 0.11, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-                    .frame(maxWidth: .infinity)
-            }
-            .frame(width: rect.width, height: rect.height)
-            .modifier(HotspotChrome(part: part, selectedPart: selectedPart, dirtyParts: dirtyParts))
-        }
-        .buttonStyle(.plain)
-        .position(x: rect.midX, y: rect.midY)
-    }
-
-    private func modeSwitchKey(width: CGFloat, height: CGFloat) -> some View {
-        let rect = frame(78.9, 40.9, 8.0, 10.2, width: width, height: height)
-        return VStack(spacing: rect.height * 0.08) {
-            ZStack {
-                RoundedRectangle(cornerRadius: rect.width * 0.2, style: .continuous)
-                    .fill(Color.white)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: rect.width * 0.2, style: .continuous)
-                            .stroke(Color.black.opacity(0.08), lineWidth: 1)
-                    )
-                    .shadow(color: .black.opacity(0.12), radius: 10, y: 4)
-
-                Image(systemName: "arrow.triangle.2.circlepath")
-                    .font(.system(size: rect.height * 0.18, weight: .semibold))
-                    .foregroundStyle(Color.black.opacity(0.42))
-            }
-            .frame(width: rect.width * 0.78, height: rect.height * 0.5)
-
-            Text("Mode")
-                .font(.system(size: rect.height * 0.1, weight: .medium))
-                .foregroundStyle(.secondary)
-                .lineLimit(1)
-        }
-        .frame(width: rect.width, height: rect.height)
-        .position(x: rect.midX, y: rect.midY)
-        .help("实体模式切换键")
-    }
-
-    private func switchButton(width: CGFloat, height: CGFloat) -> some View {
-        let part = AhaKeyStudioPart.toggleSwitch
-        let rect = frame(87.8, 35.6, 6.8, 10.6, width: width, height: height)
-        return Button {
-            onSelect(part)
-        } label: {
-            VStack(spacing: 6) {
-                ZStack(alignment: .top) {
-                    RoundedRectangle(cornerRadius: rect.width * 0.18, style: .continuous)
-                        .fill(Color.black.opacity(0.12))
-                        .overlay(
-                            RoundedRectangle(cornerRadius: rect.width * 0.18, style: .continuous)
-                                .stroke(Color.black.opacity(0.12), lineWidth: 1)
-                        )
-                    Capsule()
-                        .fill(Color.white)
-                        .frame(width: rect.width * 0.36, height: rect.height * 0.65)
-                        .overlay(Circle().fill(Color.gray.opacity(0.24)).frame(width: rect.width * 0.28, height: rect.width * 0.28))
-                        .offset(y: switchTitle == "自动批准" ? -rect.height * 0.08 : rect.height * 0.12)
-                }
-                .frame(width: rect.width * 0.58, height: rect.height * 0.78)
-
-                Text(switchTitle)
-                    .font(.system(size: rect.height * 0.12, weight: .medium))
-                    .foregroundStyle(.secondary)
-                    .lineLimit(1)
-            }
-            .frame(width: rect.width, height: rect.height)
-            .modifier(HotspotChrome(part: part, selectedPart: selectedPart, dirtyParts: dirtyParts))
-        }
-        .buttonStyle(.plain)
-        .position(x: rect.midX, y: rect.midY)
-    }
-
-    private func frame(_ x: CGFloat, _ y: CGFloat, _ w: CGFloat, _ h: CGFloat, width: CGFloat, height: CGFloat) -> CGRect {
-        CGRect(
-            x: x / baseWidth * width,
-            y: y / baseHeight * height,
-            width: w / baseWidth * width,
-            height: h / baseHeight * height
-        )
-    }
-
-    private func position(_ x: CGFloat, _ y: CGFloat, width: CGFloat, height: CGFloat) -> CGPoint {
-        CGPoint(x: x / baseWidth * width, y: y / baseHeight * height)
-    }
-
-    private func scaled(_ value: CGFloat, in width: CGFloat) -> CGFloat {
-        value / baseWidth * width
-    }
-
-    private func ledColor(for index: Int) -> Color {
-        switch AhaKeyLightBarDraft.hardwareEffect(for: lightBarPreview) {
-        case .middleLight:
-            [Color.red.opacity(0.28), Color.red.opacity(0.86), Color.red.opacity(0.86), Color.red.opacity(0.28)][index]
-        case .singleMove:
-            [Color.cyan, Color.blue.opacity(0.65), Color.cyan, Color.blue.opacity(0.65)][index]
-        case .breathing:
-            [Color.orange.opacity(0.65), Color.yellow.opacity(0.82), Color.orange.opacity(0.65), Color.yellow.opacity(0.82)][index]
-        case .rainbowMove:
-            [Color.orange, Color.yellow, Color.green, Color.blue][index]
-        case .rainbowWave:
-            [Color.pink, Color.orange, Color.green, Color.blue][index]
-        case .rainbowWaveSlow:
-            [Color.purple, Color.pink, Color.teal, Color.blue.opacity(0.8)][index]
-        case .off:
-            [Color.gray.opacity(0.18), Color.gray.opacity(0.18), Color.gray.opacity(0.18), Color.gray.opacity(0.18)][index]
-        }
-    }
-
-    private func openNativeSpeechPrivacySettings() {
-        openStudioNativeSpeechPrivacySettingsURL()
-    }
-}
-
-private func openStudioNativeSpeechPrivacySettingsURL() {
-    let candidates = [
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_Microphone",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_SpeechRecognition",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy"
-    ]
-
-    for candidate in candidates {
-        if let url = URL(string: candidate), NSWorkspace.shared.open(url) {
-            break
-        }
-    }
-}
-
-/// 输入监控 / 辅助功能 / 麦克风和语音转写：系统在「已拒绝」或部分版本下不会再弹权限窗。主动申请后打开「隐私与安全性」相关页，保证有可操作反馈。
-@MainActor
-private func openStudioCombinedVoicePrivacySettingsURL() {
-    // 勿用未文档化的 `x-apple.systemsettings` + `.extension` 等组合；在部分系统上会被当成「文稿」，
-    // 连续弹出「在 App Store 搜索… / 选取应用程序」而非进入设置。
-    let candidates = [
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_ListenEvent",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_Accessibility",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_Microphone",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy_SpeechRecognition",
-        "x-apple.systempreferences:com.apple.preference.security?Privacy",
-    ]
-    for candidate in candidates {
-        guard let url = URL(string: candidate) else { continue }
-        if NSWorkspace.shared.open(url) {
-            return
-        }
-    }
-    let appPaths = [
-        "/System/Applications/System Settings.app",
-        "/System/Library/CoreServices/Applications/System Settings.app",
-        "/System/Applications/System Preferences.app",
-    ]
-    for path in appPaths where FileManager.default.fileExists(atPath: path) {
-        if NSWorkspace.shared.open(URL(fileURLWithPath: path)) {
-            return
-        }
-    }
-}
-
-/// 先走系统 API 申请；随后在桌面端打开「隐私与安全性」相关页。输入监控 / 辅助功能在多数 macOS 版本上**不会**像 iOS 那样弹窗，麦克风和语音在「已选择过」后也不再弹窗，因此必须配合系统设置界面。
-@MainActor
-private func requestStudioPermissionsThenOpenPrivacySettingsIfNeeded(
-    voiceRelay: VoiceRelayService,
-    nativeSpeech: NativeSpeechTranscriptionService,
-    delay: TimeInterval = 0.45
-) {
-    voiceRelay.refreshPermissions(requestIfNeeded: true)
-    nativeSpeech.refreshPermissions(requestIfNeeded: true)
-    DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
-        openStudioCombinedVoicePrivacySettingsURL()
-    }
-}
-
-/// 退出后由 `open -n` 再拉起同一份 .app。须在子进程成功执行 `open` 之后再 `terminate`，否则主进程先退出会导致排队的 `open` 来不及运行。
-private func relaunchApplicationForPermissionRefresh() {
-    let bundlePath = Bundle.main.bundlePath
-    DispatchQueue.global(qos: .userInitiated).async {
-        let process = Process()
-        process.executableURL = URL(fileURLWithPath: "/usr/bin/open")
-        process.arguments = ["-n", bundlePath]
-        do {
-            try process.run()
-            process.waitUntilExit()
-        } catch {
-            // 仍尝试退出，避免卡死；用户可手动再开。
-        }
-        DispatchQueue.main.async {
-            NSApp.terminate(nil)
-        }
-    }
-}
-
-/// 在系统「隐私与安全性」中改完权限后，用确认框引导用户：退出后由 `open -n` 自动拉起同一份 .app。
-private struct RestartToApplyPermissionsButton: View {
-    var title: String = "退出并重新打开…"
-    @State private var showConfirm = false
-
-    var body: some View {
-        Button(title) { showConfirm = true }
-            .buttonStyle(.bordered)
-            .help("在系统设置中修改权限后，需重启本应用，检测才会与系统一致。")
-            .alert("需要重启以刷新权限", isPresented: $showConfirm) {
-                Button("取消", role: .cancel) {}
-                Button("立即重启") { relaunchApplicationForPermissionRefresh() }
-            } message: {
-                Text("将先退出本应用，再自动重新打开。重新打开后「重新检查权限」会读取最新系统状态。")
-            }
-    }
-}
-
-private struct HotspotChrome: ViewModifier {
-    let part: AhaKeyStudioPart
-    let selectedPart: AhaKeyStudioPart
-    let dirtyParts: Set<AhaKeyStudioPart>
-
-    func body(content: Content) -> some View {
-        content
-            .overlay(
-                RoundedRectangle(cornerRadius: 14, style: .continuous)
-                    .strokeBorder(selectedPart == part ? Color.accentColor : Color.black.opacity(0.05), lineWidth: selectedPart == part ? 2 : 1)
-            )
-            .overlay(alignment: .topTrailing) {
-                if dirtyParts.contains(part) {
-                    Circle()
-                        .fill(Color.orange)
-                        .frame(width: 8, height: 8)
-                        .padding(8)
-                }
-            }
-            .shadow(color: selectedPart == part ? Color.accentColor.opacity(0.18) : .clear, radius: 10)
-    }
-}
-
-private struct OLEDMotionPreviewSheet: View {
-    let modeTitle: String
-    let assetPath: String?
-    @Environment(\.dismiss) private var dismiss
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 16) {
-            HStack {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("\(modeTitle) 动图预览")
-                        .font(.system(size: 20, weight: .semibold))
-                    Text("这里展示的是你刚选中的 GIF 动图文件。")
-                        .font(.callout)
-                        .foregroundStyle(.secondary)
-                }
-                Spacer()
-                Button("关闭") {
-                    dismiss()
-                }
-                .buttonStyle(.bordered)
-            }
-
-            ZStack {
-                RoundedRectangle(cornerRadius: 16, style: .continuous)
-                    .fill(Color.black.opacity(0.92))
-
-                if let assetPath {
-                    DraggableAnimatedGIFPreview(path: assetPath)
-                        .padding(12)
-                } else {
-                    ContentUnavailableView("还没有选择动图", systemImage: "film.stack")
-                        .frame(minWidth: 480, minHeight: 240)
-                }
-            }
-            .frame(maxWidth: .infinity, minHeight: 240, maxHeight: 460)
-            .clipped()
-        }
-        .padding(24)
-        .frame(minWidth: 620, minHeight: 380)
-    }
-}
-
-/// 支持鼠标按住拖拽（上下左右）查看大图，避免仅靠滚轮导致横向浏览困难。
-private struct DraggableAnimatedGIFPreview: View {
-    let path: String
-    @State private var imageSize = CGSize(width: 480, height: 240)
-    @State private var offset: CGSize = .zero
-    @State private var dragStartOffset: CGSize = .zero
-
-    var body: some View {
-        GeometryReader { proxy in
-            let viewportSize = proxy.size
-            AnimatedGIFPreview(path: path)
-                .frame(width: imageSize.width, height: imageSize.height)
-                .position(
-                    x: viewportSize.width / 2 + offset.width,
-                    y: viewportSize.height / 2 + offset.height
-                )
-                .contentShape(Rectangle())
-                .gesture(
-                    DragGesture()
-                        .onChanged { value in
-                            let proposed = CGSize(
-                                width: dragStartOffset.width + value.translation.width,
-                                height: dragStartOffset.height + value.translation.height
-                            )
-                            offset = clampOffset(proposed, imageSize: imageSize, viewportSize: viewportSize)
-                        }
-                        .onEnded { _ in
-                            dragStartOffset = offset
-                        }
-                )
-                .onAppear {
-                    reloadImageSizeAndResetOffset()
-                }
-                .onChange(of: path) { _, _ in
-                    reloadImageSizeAndResetOffset()
-                }
-        }
-    }
-
-    private func reloadImageSizeAndResetOffset() {
-        if let image = NSImage(contentsOfFile: path), image.size.width > 0, image.size.height > 0 {
-            imageSize = image.size
-        } else {
-            imageSize = CGSize(width: 480, height: 240)
-        }
-        offset = .zero
-        dragStartOffset = .zero
-    }
-
-    private func clampOffset(_ proposed: CGSize, imageSize: CGSize, viewportSize: CGSize) -> CGSize {
-        let maxX = max(0, (imageSize.width - viewportSize.width) / 2)
-        let maxY = max(0, (imageSize.height - viewportSize.height) / 2)
-        return CGSize(
-            width: min(max(proposed.width, -maxX), maxX),
-            height: min(max(proposed.height, -maxY), maxY)
-        )
-    }
-}
-
-private struct AnimatedGIFPreview: NSViewRepresentable {
-    let path: String
-
-    func makeNSView(context: Context) -> NSImageView {
-        let imageView = NSImageView()
-        imageView.imageScaling = .scaleProportionallyUpOrDown
-        imageView.animates = true
-        imageView.wantsLayer = true
-        imageView.layer?.cornerRadius = 12
-        imageView.layer?.masksToBounds = true
-        imageView.imageAlignment = .alignCenter
-        return imageView
-    }
-
-    func updateNSView(_ nsView: NSImageView, context: Context) {
-        nsView.image = NSImage(contentsOfFile: path)
-    }
-}
