+++ platforms/macos/client/Sources/Views/App/AhaKeyRootWorkspaceView.swift
