+++ platforms/macos/client/Sources/Views/Device/DeviceInfoView.swift
