+++ platforms/macos/client/Sources/Views/Device/KeyMappingView.swift
