+++ platforms/macos/client/Sources/Views/Device/OLEDManagerView.swift
