+++ platforms/macos/client/Sources/Views/Studio/Sheets/AhaKeyOLEDPreviewViews.swift
@@ -1,0 +1,121 @@
+import AppKit
+import SwiftUI
+
+struct OLEDMotionPreviewSheet: View {
+    let modeTitle: String
+    let assetPath: String?
+    @Environment(\.dismiss) private var dismiss
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16) {
+            HStack {
+                VStack(alignment: .leading, spacing: 4) {
+                    Text("\(modeTitle) 动图预览")
+                        .font(.system(size: 20, weight: .semibold))
+                    Text("这里展示的是你刚选中的 GIF 动图文件。")
+                        .font(.callout)
+                        .foregroundStyle(.secondary)
+                }
+                Spacer()
+                Button("关闭") {
+                    dismiss()
+                }
+                .buttonStyle(.bordered)
+            }
+
+            ZStack {
+                RoundedRectangle(cornerRadius: 16, style: .continuous)
+                    .fill(Color.black.opacity(0.92))
+
+                if let assetPath {
+                    DraggableAnimatedGIFPreview(path: assetPath)
+                        .padding(12)
+                } else {
+                    ContentUnavailableView("还没有选择动图", systemImage: "film.stack")
+                        .frame(minWidth: 480, minHeight: 240)
+                }
+            }
+            .frame(maxWidth: .infinity, minHeight: 240, maxHeight: 460)
+            .clipped()
+        }
+        .padding(24)
+        .frame(minWidth: 620, minHeight: 380)
+    }
+}
+
+private struct DraggableAnimatedGIFPreview: View {
+    let path: String
+    @State private var imageSize = CGSize(width: 480, height: 240)
+    @State private var offset: CGSize = .zero
+    @State private var dragStartOffset: CGSize = .zero
+
+    var body: some View {
+        GeometryReader { proxy in
+            let viewportSize = proxy.size
+            AnimatedGIFPreview(path: path)
+                .frame(width: imageSize.width, height: imageSize.height)
+                .position(
+                    x: viewportSize.width / 2 + offset.width,
+                    y: viewportSize.height / 2 + offset.height
+                )
+                .contentShape(Rectangle())
+                .gesture(
+                    DragGesture()
+                        .onChanged { value in
+                            let proposed = CGSize(
+                                width: dragStartOffset.width + value.translation.width,
+                                height: dragStartOffset.height + value.translation.height
+                            )
+                            offset = clampOffset(proposed, imageSize: imageSize, viewportSize: viewportSize)
+                        }
+                        .onEnded { _ in
+                            dragStartOffset = offset
+                        }
+                )
+                .onAppear {
+                    reloadImageSizeAndResetOffset()
+                }
+                .onChange(of: path) { _, _ in
+                    reloadImageSizeAndResetOffset()
+                }
+        }
+    }
+
+    private func reloadImageSizeAndResetOffset() {
+        if let image = NSImage(contentsOfFile: path), image.size.width > 0, image.size.height > 0 {
+            imageSize = image.size
+        } else {
+            imageSize = CGSize(width: 480, height: 240)
+        }
+        offset = .zero
+        dragStartOffset = .zero
+    }
+
+    private func clampOffset(_ proposed: CGSize, imageSize: CGSize, viewportSize: CGSize) -> CGSize {
+        let maxX = max(0, (imageSize.width - viewportSize.width) / 2)
+        let maxY = max(0, (imageSize.height - viewportSize.height) / 2)
+        return CGSize(
+            width: min(max(proposed.width, -maxX), maxX),
+            height: min(max(proposed.height, -maxY), maxY)
+        )
+    }
+}
+
+private struct AnimatedGIFPreview: NSViewRepresentable {
+    let path: String
+
+    func makeNSView(context: Context) -> NSImageView {
+        let imageView = NSImageView()
+        imageView.imageScaling = .scaleProportionallyUpOrDown
+        imageView.animates = true
+        imageView.wantsLayer = true
+        imageView.layer?.cornerRadius = 12
+        imageView.layer?.masksToBounds = true
+        imageView.imageAlignment = .alignCenter
+        return imageView
+    }
+
+    func updateNSView(_ nsView: NSImageView, context: Context) {
+        nsView.image = NSImage(contentsOfFile: path)
+    }
+}
