+++ platforms/macos/client/Sources/Views/Voice/LLMConfigView.swift
