+++ platforms/macos/client/Sources/Views/Voice/VoiceAgentWorkspaceView.swift
