+++ platforms/macos/client/Sources/Views/Voice/VoiceInputFloatingHUD.swift
