+++ platforms/macos/client/Sources/Views/Workbench/AhaKeyWorkbenchView.swift
